% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_inf.R
\name{acc_inf}
\alias{acc_inf}
\title{Interpolate accurate inflection points}
\usage{
acc_inf(inds = seq_along(xvals), xvals, yvals, sd, linfs, rinfs)
}
\arguments{
\item{inds}{Optional \code{numeric} vector of data indices (by default, the sequence along retention time data is used, i.e. \code{seq_along(xvals)}).}

\item{xvals}{Retention time (x-axis) data as a \code{numeric} vector.}

\item{yvals}{Signal (y-axis) data as a \code{numeric} vector.}

\item{sd}{Second derivative of \code{yvals} as a \code{numeric} vector.}

\item{linfs, rinfs}{Indices of \strong{left} (\code{linfs}) and \strong{right} (\code{rinfs}) inflection points (e.g. approximated by \code{\link{chrom_detect}}).}
}
\value{
A \code{list} of two \code{data.frame} objects containing \code{left} and \code{right} inflection point coordinates (\code{"acc_x"}
and \code{acc_y}) as well as the data indices between which these are situated (\code{"low"} and \code{"high"}).
}
\description{
Interpolates the exact xy-coordinates (i.e. retention time and signal) of inflection point from \code{\link{chrom_detect}} approximation
by surveying a neighbourhood of nearby points.
}
\examples{
#Get retention time and signal data
rtvec <- lcqc::simlc1[,"Time"]
sigvec <- lcqc::simlc1[,"Signal"]

#Get second derivatives
sdvec <- chrom_deriv(rtvec,sigvec)[[2]]

#Set locations of left and right inflection points
lvec <- c(527,601,621,810,974,1822,2483)
rvec <- c(546,611,638,828,994,1850,2522)

acc_inf(xvals = rtvec, yvals = sigvec, sd = sdvec, linfs = lvec, rinfs = rvec)

}
\seealso{
\code{\link{acc_max}}, \code{\link{dtprep}}
}
