% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_max.R
\name{acc_max}
\alias{acc_max}
\title{Obtain accurate peak apex data via quadratic fitting}
\usage{
acc_max(inds = seq_along(xvals), xvals, yvals, maxes, linfs, rinfs, ptypes)
}
\arguments{
\item{inds}{Optional \code{numeric} vector of data indices (by default, the sequence along retention time data is used, i.e. \code{seq_along(xvals)}).}

\item{xvals}{Retention time (x-axis) data as a \code{numeric} vector.}

\item{yvals}{Signal (y-axis) data as a \code{numeric} vector.}

\item{maxes}{Indices of peak maxima (apices) as a \code{numeric} vector.}

\item{linfs, rinfs}{Indices of \strong{left} (\code{linfs}) and \strong{right} (\code{rinfs}) inflection points (e.g. approximated by \code{\link{chrom_detect}}).}

\item{ptypes}{A \code{character} vector of peak apex types. There are 4 possible values: baseline-resolved (\code{"B"}), fused (\code{"F"}),
shoulder (\code{"S"}), and round (\code{"R"}) peaks.}
}
\value{
A 3-column \code{data.frame} containing the accurate xy-coordinates and the utilized method for each peak (one of \code{"5_point"},
\code{"3_point"}, or \code{"no_fit"}).
}
\description{
Evaluates the shape of peak tops and, if appropriate, quadratically fits a parabola to derive accurate retention time (and signal) data.
}
\details{
The function tests various conditions to pick the appropriate method of estimating accurate peak retention time (and signal) among a
5-point quadratic fit or a 3-point quadratic fit. No fit is carried out for shoulder and round peaks, or those where at least one
inflection point is above the maximum in terms of signal. For others, provided the inflection point width is >5 points, a 5-point
quadratic fit is attempted. Else, or if the estimated retention time is outside that of the 5 points used for the fit, a 3-point fit is
instead attempted. If the estimated retention time is again outside that of the 3 points used for the fit, no fit is carried out.
}
\examples{
#Get retention time and signal data
rtvec <- lcqc::simlc1[,"Time"]
sigvec <- lcqc::simlc1[,"Signal"]

#Set locations of left and right inflection points as well as apices
lvec <- c(527,601,621,810,974,1822,2483)
rvec <- c(546,611,638,828,994,1850,2522)
mvec <- c(536,608,628,819,983,1834,2499)
ptype <- c("F","S","F","B","B","B","B")

#Get accurate maxima
res <- acc_max(xvals = rtvec, yvals = sigvec, maxes = mvec,
linfs = lvec, rinfs = rvec, ptypes = ptype)

}
\seealso{
\code{\link{dtprep}}, \code{\link{acc_inf}}
}
