% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_detect.R
\name{addnms}
\alias{addnms}
\title{Add peak names to \code{\link{chrom_detect}} output}
\usage{
addnms(x, nms, whichpks = NA)
}
\arguments{
\item{x}{The output of \code{\link{chrom_detect}} to add peak names to.}

\item{nms}{A \code{character} vector of peak names. Must correspond to peak indices given in \code{whichpks} (if provided).}

\item{whichpks}{An \strong{optional} \code{numeric} vector of peak indices to include names for. Must be within the range
of indices included in \code{x}. Defaults to \code{NA}, in which case \code{nms} must be provided for all peaks.}
}
\value{
A list of the same structure as that output by \code{\link{chrom_detect}} but including peak names.
}
\description{
A helper function that adds peak names to existing output of \code{\link{chrom_detect}}.
}
\examples{
\dontrun{
dt <- lcqc:::wf_detpeaks
pnms <- paste0("Ex",seq(7))
res <- addnms(dt,pnms)
}
}
