% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{apexbnds}
\alias{apexbnds}
\title{Expand baseline outward using the ApexTrack algorithm}
\usage{
apexbnds(
  end_l,
  str_l,
  str_r,
  end_r,
  rtime,
  sig,
  fder,
  sder,
  liftoff = 0,
  touchdown = 0.5,
  confirm = 0
)
}
\arguments{
\item{end_l, end_r}{Vectors of \code{numeric} indices with maximum leftmost (\code{_l}) and rightmost (\code{_r}) expansion limits (as data indices).}

\item{str_l, str_r}{Vectors of \code{numeric} indices with leftward (\code{_l}) and rightward (\code{_r}) expansion start indices (usually peak inflection points).}

\item{rtime}{A \code{numeric} vector of retention time data (x-axis).}

\item{sig}{A \code{numeric} vector of signal data (y-axis) corresponding to \code{rtime}.}

\item{fder, sder}{Both \code{numeric} vectors of first (\code{fder}) and second (\code{sder}) derivatives of \code{sig}.}

\item{liftoff, touchdown}{Single percentages (0 to 100) used as criteria to terminate leftward (\code{liftoff}) or rightward (\code{touchdown}) baseline expansion.
The cornerstone of the ApexTrack algorithm. Consult \strong{References} for further details.}

\item{confirm}{Positive \code{integer} denoting the number of consecutive points surveyed to confirm baseline expansion termination
after \code{liftoff} and \code{touchdown} criteria are satisfied.}
}
\value{
A list of length 5 with the following elements:
\describe{
\item{initial}{A \code{data.frame} containing the starting points and maximum expansion limits as given in \code{end_l}, \code{str_l}, \code{str_r}, and \code{end_r}.}
\item{starts}{A \code{numeric} vector of peak \strong{start} indices determined by ApexTrak expansion.}
\item{ends}{A \code{numeric} vector of peak \strong{end} indices determined by ApexTrak expansion.}
\item{slopes}{A \code{numeric} vector of final baseline slopes obtained after expansion.}
\item{intercepts}{A \code{numeric} vector of y-intercepts corresponding to \code{$slopes}.}
}
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Implements the Water Corporation ApexTrack baseline expansion algorithm to determine start and end boundaries of chromatographic peaks.
See \strong{Details} and \strong{References} for detailed information.
}
\details{
The ApexTrack algorithm published by Waters Corporation (updated in 2017) performs a geometric search for a baseline that best characterizes peak
starting and ending points. First, the inflection points of a peak apex are detected, an initial baseline is drawn between them and characterized by
a slope. Slope differences between this value and slopes of inflection point tangent lines are then calculated. Percentages of this difference,
determined by arguments \code{liftoff} and \code{touchdown} for the left and right peak boundaries, respectively, are set as the final criteria to
terminate baseline expansion. Once the baseline is expanded, the final slope and intercept are recorded to use for subsequent detection of crossing
baselines characteristic of \strong{fused peaks} (see \code{\link{find_fused}} and \code{\link{base_fused}}).
}
\references{
Waters Corporation (2017), 'Empower Software Data Acquisition and Processing Theory Guide', document 715005481 (Rev. A), available at: \url{https://support.waters.com/KB_Inf/Empower_Breeze/WKB57375_Empower_3_-_How_to_acquire_and_process_data (accessed 19.04.2024)}.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{base_raw}}, \code{\link{find_fused}}, \code{\link{base_fused}}
}
\keyword{internal}
