% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{base_fused}
\alias{base_fused}
\title{Fuse and expand intersecting baselines via ApexTrack}
\usage{
base_fused(
  fused,
  inds_fused,
  base_str,
  base_end,
  rtime,
  sig,
  fder,
  sder,
  liftoff,
  touchdown
)
}
\arguments{
\item{fused}{\strong{Data} indices of fused peak boundaries}

\item{inds_fused}{\strong{Peak} indices of fused peaks corresponding to \code{fused} (determined by \code{\link{find_fused}}).}

\item{base_str, base_end}{Indices of baseline-resolved peak starts and ends as determined by \code{\link{base_raw}} or another ApexTrack-based workflow.}

\item{rtime}{A \code{numeric} vector of retention time (x-axis) data.}

\item{sig}{A \code{numeric} vector of signal (y-axis) data.}

\item{fder, sder}{Both \code{numeric} vectors of first (\code{fder}) and second (\code{sder}) derivatives of \code{sig}.}

\item{liftoff, touchdown}{Single \code{numeric} parameters of ApexTrack expansion (see \code{\link{apexbnds}}).}
}
\value{
A \code{list} with 2 elements containing the final start (element \code{starts}) and end (element \code{ends})
indices of baseline-resolved peaks along \code{sig}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

This function is designed to be applied after \code{\link{base_raw}} and \code{\link{find_fused}} to fuse intersecting initial ApexTrack
baselines and carry out secondary expansion (where appropriate) and derive final start/end indices of baseline-resolved peak/peak groups.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{apexbnds}}, \code{\link{base_raw}}, \code{\link{find_fused}}
}
\keyword{internal}
