% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{base_raw}
\alias{base_raw}
\title{Derive initial expanded baselines for a series of peaks via ApexTrack}
\usage{
base_raw(rtime, sig, fder, sder, infs, sdmin, sigmax, fdcross, fdcross_nc)
}
\arguments{
\item{rtime}{A \code{numeric} vector of retention time (x-axis) data.}

\item{sig}{Signal (y-axis) data corresponding to \code{rtime}.}

\item{fder, sder}{First and second derivatives of \code{sig}, e.g. obtained via \code{\link{chrom_deriv}}.}

\item{infs}{A \code{list} of length 2 containing \strong{left} and \strong{right} inflection points as separate elements.}

\item{sdmin}{Indices of second derivative minima along \code{sig}.}

\item{sigmax}{Indices of signal maxima along \code{sig}.}

\item{fdcross, fdcross_nc}{Indices of compliant and non-compliant (\code{_nc}) first derivative downcrosses.}
}
\value{
Baseline-resolved peak start and end indices alongside other data structured identically to the usual output of \code{\link{apexbnds}}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

The function first filters all second derivative minima (\code{sdmin}) which are likely to represent shoulder or round peaks.
These include peaks which have less than 2 proximal inflection points \strong{and} do not have corresponding signal maxima.
ApexTrack baseline expansion (see \code{\link{apexbnds}}) is then carried out on the remaining peaks to derive initial estimates of start and end
indices for baseline-resolved peaks.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{apexbnds}}, \code{\link{find_fused}}, \code{\link{base_fused}}
}
\keyword{internal}
