% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_visc.R
\name{bilin_interp}
\alias{bilin_interp}
\title{Bilinear Interpolation of Irregularly-spaced Data}
\usage{
bilin_interp(x, y, z, xout, yout)
}
\arguments{
\item{x, y}{Vectors (\code{numeric}) of x- and y-coordinates.}

\item{z}{A vector of \code{numeric} values to populate the a \code{matrix} of dimensions \code{[x,y]}.}

\item{xout, yout}{Values of x- and y-coordinates to interpolate for.}
}
\value{
A \code{list} of vectors containing the interpolation coordinates (\code{"x"}, \code{"y"}) and interpolated values (\code{"z"}).
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_visc}} workflow, not intended for standalone use.
Carries out bilinear interpolation of irregularly-spaced data. Modified from \code{\link[akima]{bilinear}} in package \pkg{akima}.
}
\seealso{
\code{\link{chrom_tplate}}
}
\keyword{internal}
