% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{bline_als}
\alias{bline_als}
\title{Iterative Asymmetric Least Squares (ALS) baseline correction}
\usage{
bline_als(
  input,
  lambda = 6,
  p = 0.001,
  prec = 1e-08,
  maxit = 50,
  rm_neg = TRUE
)
}
\arguments{
\item{input}{Numeric vector of signal to be baseline-corrected.}

\item{lambda}{Second derivative constraint.}

\item{p}{Weights of residuals (positive).}

\item{prec}{Precision (error) tolerance required for baseline correction to reach convergence.}

\item{maxit}{Maximum number of iterations (defaults to \code{50}).}

\item{rm_neg}{A \code{TRUE/FALSE logical} specifying whether points where the computed baseline is higher that the original signal should be discarded (\code{TRUE} by default).}
}
\value{
A list with 5 elements: \code{Original_Signal}, \code{Corrected_Signal}, \code{Baseline}, \code{Method} (always \code{"als"}), and \code{Parameters}.
}
\description{
Asymmetric Least Squares (ALS) baseline correction, modified from package \pkg{baseline} for iterative convergence based on weighed Eilers smoothing (see also \code{\link{smooth_whit}}).
}
\examples{
res <- bline_als(lcqc::exgc1[,"Signal"])
bline_plot(res)

}
\references{
Peng, J., Peng, S., Jiang, A., Wei, J., Li, C., Tan, J. (2010), 'Asymmetric Least Squares for Multiple Spectra Baseline Correction', \emph{Analytica Chimica Acta} \strong{683} (1), pp. 63-68, doi: \href{10.1016/j.aca.2010.08.033}{https://www.doi.org/10.1016/j.aca.2010.08.033}.
}
\seealso{
\code{\link{chrom_bline}}, \code{\link{smooth_whit}}
}
