% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{bline_chang}
\alias{bline_chang}
\title{Baseline correction (Chang's method)}
\usage{
bline_chang(
  input,
  threshold = 0.5,
  alpha = 0.95,
  bfrac = 0.2,
  segments = 100,
  sig_window = 10,
  fit = "linear",
  rm_neg = TRUE
)
}
\arguments{
\item{input}{Input numeric vector containing signal to be baseline-corrected.}

\item{threshold}{Position of the baseline relative to the noise component of the signal. Must be between 0 and 1.}

\item{alpha}{High-pass filter parameter.}

\item{bfrac}{Fraction of low-intensity fragments of the filtered signal that are assumed to be baseline.}

\item{segments}{Number of segments to divide the filtered signal into.}

\item{sig_window}{Signal window size (in points).}

\item{fit}{Method used for baseline fitting. One of \code{"linear"} or the significantly slower \code{"spline"} (cubic).}

\item{rm_neg}{A \code{TRUE/FALSE logical} specifying whether points where the computed baseline is higher that the original signal should be discarded (\code{TRUE} by default).}
}
\value{
A list with 5 elements: \code{Original_Signal}, \code{Corrected_Signal}, \code{Baseline}, \code{Method} (always \code{"chang"}), and \code{Parameters}.
}
\description{
Implementation of baseline correction according to Chang et al. (2007). Modified from package \pkg{TargetSearch}.
}
\examples{
res <- bline_chang(lcqc::exgc1[,"Signal"])
bline_plot(res)

}
\references{
Chang, D., Banack, C.D., Shah, S.L. (2007), 'Robust baseline correction algorithm for signal dense NMR spectra', \emph{Journal of Magnetic Resonance} \strong{187}, pp. 288-292.
}
\seealso{
\code{\link{chrom_bline}}
}
