% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{bline_isrea}
\alias{bline_isrea}
\title{Iterative Smoothing-Splines with Root Error Adjustment (ISREA) baseline correction}
\usage{
bline_isrea(input, eta = 10, maxit = 100, rm_neg = TRUE)
}
\arguments{
\item{input}{Input numeric vector containing signal to be baseline-corrected.}

\item{eta}{Convergence criterion. May take values from 0.0001 to 10 (based on Xu et al., 2021). Usually does not appreciably affect the results, but more computation time is needed for lower values.}

\item{maxit}{Maximum number of iterations (defaults to \code{100}).}

\item{rm_neg}{A \code{TRUE/FALSE logical} specifying whether points where the computed baseline is higher that the original signal should be discarded (\code{TRUE} by default).}
}
\value{
A list with 5 elements: \code{Original_Signal}, \code{Corrected_Signal}, \code{Baseline}, \code{Method} (always \code{"isrea"}), and \code{Parameters}.
}
\description{
Baseline correction that uses smoothing splines to estimate the baseline. Developed by Xu et al. (2021).
}
\examples{
res <- bline_isrea(lcqc::exgc1[,"Signal"])
bline_plot(res)

}
\references{
Xu, Y., Du, P., Senger, R., Robertson, J., Pirkle, J.L. (2021), 'ISREA: An Efficient Peak-Preserving Baseline Correction Algorithm for Raman Spectra', \emph{Applied Spectroscopy} \strong{75} (1), pp. 34-45, doi: \href{10.1177/0003702820955245}{https://www.doi.org/10.1177/0003702820955245}.
}
\seealso{
\code{\link{chrom_bline}}
}
