% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{bline_plot}
\alias{bline_plot}
\title{Plot baseline correction results}
\usage{
bline_plot(
  input,
  xvar = "auto",
  xname = "Index",
  bline = FALSE,
  corrsig = TRUE,
  asprat = 0.71
)
}
\arguments{
\item{input}{Output from one of the baseline correction functions of \pkg{lcqc}: \code{\link{bline_als}}, \code{\link{bline_chang}}, \code{\link{bline_poly}}, \code{\link{bline_isrea}}, or \code{\link{chrom_bline}}.}

\item{xvar}{An optional numeric vector of \code{x} values corresponding to the signal in \code{input}. If set to \code{"auto"}, uses point indices instead.}

\item{xname}{Title of x-axis (defaults to \code{"Index"}).}

\item{bline}{A \code{logical}. Should the calculated baseline be included in the plot? Defaults to \code{FALSE}.}

\item{corrsig}{A \code{logical}. Should be baseline-corrected signal be included in the plot? Defaults to \code{TRUE}.}

\item{asprat}{A numeric value setting the plot aspect ratio.}
}
\value{
A \code{ggplot}-class object containing the plot.
}
\description{
Create a summary plot of baseline correction results containing the original signal, calculated baseline, and (optionally) baseline-corrected signal.
}
\examples{
bline_res <- chrom_bline(lcqc::exgc1[,"Signal"], method = "als", slnt = TRUE)
bline_plot(bline_res, bline = TRUE, corrsig = TRUE)

}
\seealso{
\code{\link{chrom_bline}}
}
