% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{bline_poly}
\alias{bline_poly}
\title{Modified Polynomial Fit (ModPolyFit) baseline correction}
\usage{
bline_poly(input, deg = 4, prec = 0.001, maxit = 100, rm_neg = TRUE)
}
\arguments{
\item{input}{Input numeric vector containing signal to be baseline-corrected.}

\item{deg}{Degree of the polynomial fit.}

\item{prec}{Precision (error) tolerance required for baseline correction to reach convergence.}

\item{maxit}{Maximum number of iterations (defaults to \code{100}).}

\item{rm_neg}{A \code{TRUE/FALSE logical} specifying whether points where the computed baseline is higher that the original signal should be discarded (\code{TRUE} by default).}
}
\value{
A list with 5 elements: \code{Original_Signal}, \code{Corrected_Signal}, \code{Baseline}, \code{Method} (always \code{"poly"}), and \code{Parameters}.
}
\description{
Iterative baseline correction algorithm based on automated polynomial fitting developed by Lieber & Mahadevan-Jansen (2003). Modified from package \pkg{baseline}.
}
\examples{
res <- bline_poly(lcqc::exgc1[,"Signal"])
bline_plot(res)

}
\references{
Lieber, C.A., Mahadevan-Jansen, A. (2003), 'Automated Method for Subtraction of Fluorescence from Biological Raman Spectra', \emph{Applied Spectroscopy} \strong{57} (11), pp. 1363-1367, doi: \href{10.1366/000370203322554518}{https://www.doi.org/10.1366/000370203322554518}.
}
\seealso{
\code{\link{chrom_bline}}
}
