% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_pack.R
\name{chk_pack}
\alias{chk_pack}
\title{Check for and install missing packages}
\usage{
chk_pack(packlist)
}
\arguments{
\item{packlist}{A list of packages to check for. Packages not found are automatically installed and called.}
}
\value{
Nothing.
}
\description{
Checks a character vector of package names, detects any CRAN packages not currently installed, installs, and loads them.
}
\examples{
\dontrun{
chk_pack(c("ggplot2","knitr"))
}
}
