% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chkdt}
\alias{chkdt}
\title{Check suitability of input data}
\usage{
chkdt(dt)
}
\arguments{
\item{dt}{Data object output from \code{chrom_detect}}
}
\value{
Nothing.
}
\description{
\strong{This function is not exported}.

Check suitability of input data (must be output from \code{chrom_detect}).
}
\keyword{internal}
