% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_addmets.R
\name{chrom_addmets}
\alias{chrom_addmets}
\title{Calculate additional HPLC column-specific performance metrics}
\usage{
chrom_addmets(
  which_mets = "all",
  t0,
  len,
  flow = NA,
  id = NA,
  deltap = NA,
  visc = NA,
  dp = NA
)
}
\arguments{
\item{which_mets}{A \code{character} vector of metrics to calculate. One or more of: \code{"all"} (default),
\code{"linvel"} (Linear Velocity; mm/s), \code{"porosity"} (Packing Porosity; dimensionless),
\code{"flowres"} (Flow Resistance; dimensionless), \code{"pabil"} (Permeability; mm^2/s/bar),
and/or \code{"spabil"} (Specific Permeability; mm^2).}

\item{t0}{The \code{numeric} value for dead time (i.e. breakthrough time) in \strong{minutes}.}

\item{len}{The \code{numeric} column length (in mm).}

\item{flow}{Flow rate (mL/min, \code{numeric}).}

\item{id}{Column internal diameter (in mm, \code{numeric}).}

\item{deltap}{Back pressure (bar; \code{numeric}).}

\item{visc}{Dynamic viscosity of the mobile phase (mPas; \code{numeric}). May be calculated via \code{\link{chrom_visc}}.}

\item{dp}{Stationary phase particle size (in µm).}
}
\value{
A named \code{list} of length 3 containing a \code{data.frame} of \code{results}, various \code{information} about them,
and the function \code{call}.
}
\description{
This function uses various column characteristics such as length, particle size,
flow rate, internal diameter (among others) to calculate various column performance indicators.
See \strong{Details} for further information.
}
\details{
The linear velocity \eqn{\mu} of the mobile phase is simply related to column length \eqn{L_c} and true breakthrough time \eqn{t_0}.
\deqn{\mu = L_c/t_0}
It may be beneficial to verify breakthrough time in this equation, i.e. determine whether the non-retained analyte is really passing
through the column at the velocity of the mobile phase, by calculating packing porosity \eqn{\epsilon}. The packing porosity of
\strong{bonded} silica-based stationary phases is \emph{ca.} 0.65. Using this value as reference, \eqn{\epsilon} may be calculated
simply from \eqn{t_0\text{ }(s)}, \eqn{L_c\text{ }(mm)}, flow rate \eqn{F\text{ }(mL\text{ }min^{-1})}, and inner column diameter
\eqn{d_c^2\text{ }(mm^2)}.
\deqn{\epsilon = 21\times[(Ft_0)/(d_c^2L_c)]}
Significantly higher or lower values of \eqn{\epsilon} (>1 and <0.5, respectively) indicate retention of the analyte or its exclusion
from the pores of the stationary phase (Meyer, 2010).
Another useful metric is permeability \eqn{K}, which incorporates back pressure \eqn{\Delta\rho\text{ }(bar)}. A large value of \eqn{K}
indicates poor column packing, while the reverse is characteristic of a leak. Values of \eqn{15\text{ }mm^2\text{ }s^{-1}\text{ }bar^{-1}}
are typical for bonded phases at a fast flow rate of \eqn{1.6\text{ }mL\text{ }min^{-1}}.
\deqn{K = L_c^2/(\Delta\rho t_0)}
Specific permeability \eqn{K^{\circ}\text{ }(mm^{-2})} may also be calculated (typically \eqn{4.0\times 10^{-8}\text{ }mm^2} for bonded
phases) by incorporating flow rate, dynamic viscosity, column length, inner diameter, and back pressure.
\deqn{K^{\circ} = 21\times 10^-8\times[(F\eta L_c)/(d_c^2\Delta\rho)]}
Finally, \eqn{K} may be represented as a dimensionless metric called Flow Resistance \eqn{\Phi}, which facilitates comparison of different columns.
One of the additional required parameters is particle size \eqn{d_p\text{ }(\mu m)}.
\deqn{\Phi = 4.7\times[(\Delta\rho d_p^2 d_c^2)/(L_c \eta F)]}.
A typical \eqn{\Phi} value of 1000 is observed for packed HPLC columns. Significant upward and downward deviations (e.g. >2000 or <500)
are indicative of a blockage or voids in the packing, respectively.

For calculation of theoretical plates and additional dimensionless metrics such as reduced plate height and Separation Impedance
to assess column performance, see \code{\link{chrom_tplate}}.
}
\examples{
chrom_addmets(t0 = 1, len = 250, id = 3.2, flow = 1.6, deltap = 70, visc = 0.33, dp = 5)

}
\references{
Meyer, V.R. (2010), \emph{Practical High-Performance Liquid Chromatography}, John Wiley & Sons, Chichester, United Kingdom.
}
\seealso{
\code{\link{chrom_visc}}, \code{\link{chrom_tplate}}
}
