% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_amplim.R
\name{chrom_amplim}
\alias{chrom_amplim}
\title{Find amplitude limit for peak detection}
\usage{
chrom_amplim(x, method = "diff", pars = 1)
}
\arguments{
\item{x}{Numeric vector of signal from which to derive an amplitude limit.}

\item{method}{One of: \code{"quant"} (simple quantiles), \code{"diff"} (relative differences between quantiles), or \code{"zscore"} (z-scores).}

\item{pars}{Parameters required for calculation depending on the chosen \code{method}:
\describe{
\item{For \code{"quant"}}{The quantile probability to use as the amplitude limit. Given as a single \code{numeric} percentage between 0.001 and 50.}
\item{For \code{"diff"}}{The percentage (between 0.001 and 50) of maximum difference between quantiles (calculated via a probability \code{seq(0, 1, 0.005)}) used to locate the amplitude limit.}
\item{For \code{"zscore"}}{A numeric vector of 3 parameters:
the \code{integer} number of observations used to calculate moving average and standard deviation (\code{lag} in \code{\link{z_thres}}),
the \code{numeric} threshold (as a factor of moving standard deviation) at which to signal the presence of a peak (\code{threshold} in \code{\link{z_thres}}),
and \code{numeric} sensitivity of the amplitude limit to the average standard deviation throughout the signal (see \strong{Details}).}
}}
}
\value{
A single \code{numeric} value of the amplitude limit
}
\description{
Detects a suitable signal amplitude limit for peak detection using a simple quantile, relative differences between quantiles, or z-scores (see \code{\link{z_thres}}).
}
\details{
When \code{method} is \code{"quant"}, the amplitude limit is simply determined as the quantile at the specified probability percentage between 0.001 and 50.

Method \code{"diff"} first calculates the quantile range of input vector \code{x} using \code{quantile(x, probs = c(0, 1, 0.005))},
derives the maximum difference between successive quantiles (via \code{max(diff())}),
locates the earliest (lowest) quantile where a percentage of this difference between 0.001 and 50 (given in the \code{pars} argument) is exceeded,
and uses this quantile as the amplitude limit.

Finally, \code{method = "zscore"} calculates \strong{global mean values} of both the moving average \eqn{\overline{MX}} and the moving standard deviation \eqn{\overline{SD}}
of the signal (derived via \code{z_thres}) and uses a sensitivity parameter \eqn{sens} (provided in \code{pars[3]}) to calculate the amplitude limit as follows:
\deqn{AmpLim = \overline{MX} + (1/sens)\times\overline{SD}}
Thus, \eqn{AmpLim} decreases with increasing \eqn{sens} (\code{pars[3]}).
}
\examples{
sig <- lcqc::simlc1[,"Signal"]
lim1 <- chrom_amplim(sig, method = "quant", pars = 0.05)
lim2 <- chrom_amplim(sig, method = "diff", pars = 0.05)
lim3 <- chrom_amplim(sig, method = "zscore", pars = c(30, 5, 2))
lim3_alt <- chrom_amplim(sig, method = "zscore", pars = c(30, 5, 10))

}
\seealso{
\code{\link{z_thres}}
}
