% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_asym.R
\name{chrom_asym}
\alias{chrom_asym}
\title{Calculate asymmetry metrics for chromatographic peaks}
\usage{
chrom_asym(
  input,
  method = "all",
  which_peaks = "all",
  show_widths = TRUE,
  crit_w = "auto",
  asprat = 0.71,
  tpa_thres = 0.85,
  optmet = "nlp",
  plotset = "make"
)
}
\arguments{
\item{input}{The output data from function \code{\link{chrom_detect}}.}

\item{method}{A \code{character} vector of method(s) to apply for asymmetry calculations. One or more of: \code{"all"} (default),
\code{"Tf"} (USP \eqn{T_f}), \code{"As"} (\eqn{A_s}), and/or \code{"TPA"} (Total Peak Analysis).}

\item{which_peaks}{Selects specific peaks from \code{input} to process. Either \code{"all"} (default) or a \code{numeric} vector
of peak indices included in \code{input}.}

\item{show_widths}{A \code{logical} specifying whether the various peak half-widths at specific heights required for the
calculations are included in the results (\code{TRUE} by default).}

\item{crit_w}{The critical width parameter to use for baseline calculation via FastChrom (\code{\link{fastchrom_bline}}).
Defaults to \code{"auto"} or can be set manually as a \code{numeric} (usually equal to the minimum peak width at half height).}

\item{asprat}{Aspect ratio of the plot(s). Defaults to \code{0.71}.}

\item{tpa_thres}{The \code{numeric} threshold between 0 and 0.99 to use for calculation of peak width and its relation to
Gaussian standard deviation for TPA. Defaults to \code{0.85}. See \strong{Details}.}

\item{optmet}{A \code{character} string specifying which method to use for iterative adjustment of the Gaussian model.
One of \code{"optim"} or \code{"nlp"} (default).}

\item{plotset}{A \code{character} string specifying whether results of TPA are visualized/shown via \code{\link{chrom_tpa}}.
One of \code{"make"} (generates plots without printing; default), \code{"print"} (generates and prints plots), or \code{"none"}.}
}
\value{
A named \code{list} of length 4 containing the \code{data.frame} of \code{results}, the function \code{call},
the list of TPA \code{plots} (if any), and a \code{character} string providing various \code{information} about the results.
}
\description{
Calculates common asymmetry metrics for chromatographic data including the United States Pharmacopoeia (USP)/European
Pharmacopoeia (EP) Tailing Factor (\eqn{T_f}), and the Asymmetry Factor (\eqn{A_s}). Additionally, the Total Peak Analysis (TPA) workflow
proposed by Wahab et al. (2017) is also implemented for baseline-resolved peaks. See \strong{Details} for further information.
}
\details{
Peak asymmetry may be described in different ways, with most of the classical values providing a single numeric metric.
All of the methods described below have been implemented in \pkg{lcqc}.
Among the widespread metrics is the Asymmetry Factor \eqn{A_s}, calculated as the ratio between peak half-widths at 10\% peak height on
the \strong{trailing} (\eqn{B_{10}}) and \strong{leading} (\eqn{A_{10}}) edges of the peak. Thus, values greater than 1 indicate tailing,
values less than one indicate fronting, and a tailing factor of exactly 1 is characteristic of a perfectly symmetrical peak.
\deqn{A_s = B_{10}/A_{10}}
Another metric commonly calculated is the USP/EP Tailing Factor \eqn{T_f}, which uses half-widths at 5\% peak height.
\deqn{T_f = \frac{(A_5+B_5)}{2A_5}}
Irrespective of whether \eqn{T_f} or \eqn{A_s} is used, acceptance criteria outlined in various pharmacopoeias list a range of \strong{0.8-1.8}
as acceptable. In practice, values of \strong{0.9-1.2} are routinely achieved and desirable for typical test analytes.
Both \eqn{T_f} and \eqn{A_s} provide information only about the \strong{relative} amount of tailing or fronting, but fail to accommodate cases
where both are present. For example, a widened peak with both tailing and fronting (a so-called \emph{Eiffel Tower} effect) may have \eqn{T_f}
and \eqn{A_s} values close to 1 despite being heavily distorted from a Gaussian profile. In order to provide separate measures of the absolute
and relative contributions of tailing and fronting to the total peak shape, as well as to effectively visualize the results, \pkg{lcqc}
also implements the Total Peak Analysis (TPA) workflow developed by Wahab et al. (2017) as a simple visual and quantitative assessment tool.
First, the Gaussian standard deviation \eqn{\sigma} is estimated from a peak (with a maximum signal \strong{normalized to unity}) using
the peak width at a chosen percentage of peak height (\eqn{W_H}):
\deqn{\sigma = W_H/(2\sqrt{2ln(1/H)})}
The estimation of \eqn{\sigma} takes advantage of the fact that the tops of chromatographic peaks (after 80-85\% peak height) usually follow a
Gaussian distribution closely even for heavily-distorted peaks. Thus, a Gaussian model is constructed using the estimated \eqn{\sigma},
a peak maximum equal to 1, and the actual peak retention time. A linear constrained solver is then used to ensure the top 15\% of the
Gaussian model values are either equal to or enclosed by the actual chromatographic peak. Finally, absolute and relative (\%) residuals are
calculated separately for the leading and trailing edges of the peak, providing separate quantitative measures of the degree of peak fronting
and tailing. Currently, the TPA procedure is limited by \pkg{lcqc} to only assess baseline-resolved peaks. Each peak is also assessed for
its suitability for TPA based on residual sums to the left and right of the peak apex. The Gaussian model is supposed to be completely enclosed
by the original chromatographic peak, but this is seldom the case, resulting in negative residuals where the model is outside the retention time
boundaries of the peak. If \strong{>50\%} of residuals are found to be negative on either side of the peak, it is considered to be unsuitable
for TPA.
}
\examples{
\dontrun{
#Get data
dt <- lcqc:::wf_detpeaks
res <- chrom_asym(dt, which_peaks = 3:7, show_widths = FALSE)
}

}
\references{
Meyer, V.R. (2010), \emph{Practical High-Performance Liquid Chromatography}, John Wiley & Sons, Chichester, United Kingdom.

The United States Pharmacopeial Convention (2021), \emph{Physical Tests/<621> Chromatography} (USP 40-NF 35), available at:
\url{https://www.usp.org/harmonization-standards/pdg/excipients/chromatography} (accessed 24.04.2024).

Wahab, M.F., Patel, D.C., Armstrong, D.W. (2017), 'Total Peak Shape Analysis: Detection and Quantitation of Concurrent Fronting, Tailing, and Their Effect on Asymmetry Measurements', \emph{Journal of Chromatogpraphy A} \strong{1509}, pp. 163-170, DOI: \url{https://doi.org/10.1016/j.chroma.2017.06.031}.
}
\seealso{
\code{\link{chrom_tpa}}, \code{\link{chrom_detect}}
}
