% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_deriv.R
\name{chrom_deriv}
\alias{chrom_deriv}
\title{Calculate first and second derivatives}
\usage{
chrom_deriv(x, y)
}
\arguments{
\item{x}{Numeric vector of \code{x} values.}

\item{y}{Numeric vector of \code{y} values equal in length to that of \code{x}.}
}
\value{
A \code{list} of two numeric vectors equal to \code{x} in length, containing first (\code{[[1]]}) and second (\code{[[2]]}) derivatives.
}
\description{
Calculate first and second derivatives from \code{x} and \code{y} data.
}
\examples{
xvals <- lcqc::simlc5[,"Time"]
yvals <- lcqc::simlc5[,"Signal"]
chrom_deriv(xvals, yvals)
}
