% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_detect.R
\name{chrom_detect}
\alias{chrom_detect}
\title{Automatic detection and classification of chromatographic peaks}
\usage{
chrom_detect(
  chrom,
  vars = c("Time", "Signal"),
  trange = range(chrom[, vars[1]]),
  det_bunch = TRUE,
  pnms = NA,
  bline_method = "none",
  bline_pars = "default",
  smooth = rep("tri", 2),
  mpts = c(3, 3),
  crosspts = c(2, 2),
  ma_pts = c("auto", 7),
  ma_passes = c("auto", 1),
  sens = c(2, 1, 3),
  amp_thres = "zscore",
  ampfrac = 0.05,
  der_thres = c("ncore", "iqr"),
  apex_pars = c(0, 0.5),
  rej = c(wd = "auto", pa = NA, sn = NA, ht = NA),
  asprat = 0.71,
  rej_logic = rep("OR", 2),
  zscore_pars = "default",
  plot_corr = FALSE,
  plotset = "make",
  fchrom = NA
)
}
\arguments{
\item{chrom}{A \code{data.frame} containing the chromatogram. Must include x- and y-axis data variables as listed in \code{vars}. When \code{NA},
these are assumed to be contained within the first two columns of the \code{data.frame}.}

\item{vars}{Variable names of \code{chrom} corresponding to the retention time (x-axis) and signal (y-axis), in that order.
Defaults to \code{c("Time","Signal")}.}

\item{trange}{A \code{numeric} vector containing the retention time range of \code{chrom} to apply the workflow to. By default, the entire
chromatogram is processed.}

\item{det_bunch}{A \code{logical} switch. Should peak bunching be detected and mitigated? Defaults to \code{TRUE}.}

\item{pnms}{\strong{Optional} \code{character} vector of peak names. Defaults to \code{NA}.
The number of peaks detected by the algorithm must be known before providing this parameter.
Thus, a "practice run" is recommended before any names are set. Names may also be added to existing \code{chrom_detect} output
via the helper function \code{\link{addnms}}.}

\item{bline_method}{Baseline correction method to use (if any). Defaults to \code{"none"}. Possible methods are:
\code{"als"}, \code{"chang"}, \code{"poly"}, or \code{"isrea"}. See \code{\link{chrom_bline}} for details.}

\item{bline_pars}{Parameters of the baseline correction. Set to \code{"default"} (sensible defaults. Otherwise, a \strong{named} \code{numeric}
vector of parameters unique to the chosen baseline correction method (\code{bline_method}). See \code{\link{chrom_bline}} for further details.}

\item{smooth}{A \code{character} vector of length 2 specifying the smoothing method to use for signal and first/second derivatives, respectively.
Defaults to \code{rep("tri",2)} (triangular smoothing for both). A single \code{character} value may also be provided and the same method is then
applied to both signal and derivatives. For further details, see \code{\link{chrom_smooth}}.}

\item{mpts}{A \code{numeric} vector of length 2 specifying the number of points to survey on either side to confirm (or reject) detected
signal (\strong{first element}) and derivative (\strong{second element}) extremes (i.e. maxima and minima). A single \code{numeric} value
may also be provided and will then be used for both data types. Defaults to \code{c(3,3)}.}

\item{crosspts}{A \code{numeric} vector of length 2 specifying the number of points to survey on either side to confirm (or reject) detected
zero crossings (downcrosses and upcrosses) for signal (\strong{first element}) and derivatives (\strong{second element}).}

\item{ma_pts, ma_passes}{The \code{numeric} vector of length 2 specifying the number of smoothing \strong{points} (\code{ma_pts})
and \strong{passes} (\code{ma_passes}) to use to calculate smoothed signal and derivatives using the method(s) given in \code{smooth}.
The first element may also be set to \code{"auto"}, in which case \code{ma_pts} and/or \code{ma_passes} will be \strong{automatically} determined
via \code{\link{chrom_width}}. The second element must then be a \strong{number} specifying the starting points/passes to use in
\code{\link{chrom_width}}. The defaults for \code{ma_pts} and \code{ma_passes} are \code{c("auto",7)} and \code{c("auto",1)}, respectively.}

\item{sens}{A \code{numeric} vector of sensitivity parameters used for fine-tuning the first and second derivative peak detection thresholds
via function \code{\link{chrom_derlims}} (first two elements, respectively), and the amplitude limit obtained via \code{\link{chrom_amplim}}
(when \code{amp_thres} includes \code{"zscore"}). Defaults to \code{c(2,1,3)}.}

\item{amp_thres}{A \code{character} value/vector specifying method(s) to use for determination of the amplitude limit for peak detection.
\strong{One or more} of: \code{"quant"}, \code{"diff"}, or \code{"zscore"} (see \code{\link{chrom_amplim}}) for details).}

\item{ampfrac}{A parameter for \code{\link{chrom_amplim}} that helps determine the amplitude threshold via method(s) specified in \code{amp_thres}.}

\item{der_thres}{A \code{character} vector of length 2 specifying the methods to use for determination of derivative-based peak detection thresholds
via \code{\link{chrom_derlims}}. The \strong{first} element must be one of \code{"vaz"}, \code{"zscore"}, or \code{"ncore"}, while possible values
for the \strong{second} element are \code{"none"}, \code{"all"}, \code{"iqr"}, \code{"quant"}, or \code{"sd"}.}

\item{apex_pars}{The \strong{liftoff} and \strong{touchdown} parameters to use for ApexTrack baseline expansion (see \strong{Details}, \strong{References}, and \code{\link{apexbnds}} for details).}

\item{rej}{A \strong{named} \code{numeric} vector specifying peak filters to apply for selective peak removal (carried out in two stages, before and after boundary detection and classification).
Possible names are:
\enumerate{
\item For \strong{pre-}classification (filtering round 1): minimum height (\code{"ht"}), S/N ratio (\code{"sn"}).
\item For \strong{post-}classification (filtering round 2): inflection point width (\code{"wd"}) and peak area (\code{"pa"}).
}
All values default to \code{NA} (no filters are applied).}

\item{asprat}{Aspect ratio of plots (defaults to \code{0.71}).}

\item{rej_logic}{A \code{character} vector of length 2 specifying the logic to use for two rounds of peak filtering (\strong{pre-} and \strong{post-}classification).
Possible values are \code{"OR"} or \code{"AND"}. Defaults to \code{rep("OR",2)}. A single value may also be provided, in which case the same logic will
be applied for both rounds of peak filtering.}

\item{zscore_pars}{Either \code{"default"} or a \code{list} of \code{numeric} vectors providing sequences of 3 parameters to use for \code{\link{z_optim}}.}

\item{plot_corr}{A \code{logical} switch. Should the baseline-corrected signal be plotted?
When \code{FALSE} (default), the original signal is plotted instead.}

\item{plotset}{A \code{character} string specifying whether data is visualized/shown.
One of \code{"make"} (generates plots without printing; default), \code{"print"} (generates and prints plots), or \code{"none"}.}

\item{fchrom}{Either \code{NA} (default) or a \code{numeric} value of critical width to use for FastChrom baseline correction.
Applies to resolved peaks only. See \code{\link{fastchrom_bline}} for details.}
}
\value{
A \strong{list of lists} containing named elements \code{results} and \code{plots}.
The latter includes 3 \code{ggplot}-class plot objects including the chromatogram with detected peak markers as well as both first and second derivative-based noise plots.
Element \code{results} contains the following elements in a \code{list}:
\describe{
\item{Chromatogram}{The chromatogram \code{data.frame} containing retention times, signal, baseline-corrected and/or smoothed signals (if any),
original and (if any) smoothed first/second derivatives, and peak ID assignments.}
\item{Derivative_Noise}{A 2-column \code{data.frame} containing the first and second derivative noise from \code{\link{chrom_derlims}}.}
\item{Peaks}{A \code{list} of \code{data.frame} objects, each containing data for an individual detected peak. This includes retention time, original and (if any) smoothed signals, and associated first/second derivatives.}
\item{Peak_Extents}{A \code{data.frame} containing key information about the detected peaks as detailed in \code{\link{peakfind}} (such as peak types, boundary types, inflection and upslope points etc.).}
\item{Amplitude_Limit}{The \code{numeric} peak detection amplitude limit.}
\item{Derivative_Limits}{A \code{list} with lower and upper peak detection thresholds derived from first and second derivatives (elements \code{FD} and \code{SD}, respectively).}
\item{Zscore_Limits}{A \code{numeric} vector of starting and ending retention times of peak regions derived from \code{\link{z_optim}}.}
\item{information}{A \code{character} string containing summary information about the results.}
\item{call}{The function call.}
}
}
\description{
This function uses a combination of various custom and industry-standard algorithms to automatically detect, filter, and classify
chromatographic peaks using 1-dimensional data (only retention time and signal). This is the most complex function in \pkg{lcqc}.
}
\examples{
#LC chromatogram (with FastChrom baseline correction)
reslc1 <- chrom_detect(lcqc::simlc1, vars = c("Time","Signal"), fchrom = 1)

#GC chromatogram
resgc <- chrom_detect(lcqc::exgc2, vars = c("Time","Signal"), det_bunch = FALSE)

#GC chromatogram with baseline correction
resgc2 <- chrom_detect(lcqc::exgc1, vars = c("Time","Signal"),
bline_method = "als", det_bunch = FALSE)

}
\references{
Pirttilä, K., Balgoma, D., Rainer, J., Pettersson, C., Hedeland, M., Brunius, C. (2022), 'Comprehensive Peak Characterisation (CPC) in Untargeted LC-MS Analysis', \emph{Metabolites} \strong{12} (2), article 137, DOI: \url{https://www.doi.org/10.3390/metabo12020137}.

Waters Corporation (2017), 'Empower Software Data Acquisition and Processing Theory Guide', document 715005481 (Rev. A), available at: \url{https://support.waters.com/KB_Inf/Empower_Breeze/WKB57375_Empower_3_-_How_to_acquire_and_process_data (accessed 19.04.2024)}.
}
\seealso{
This workflow uses a multitude of exported and \strong{un}exported functions:
\describe{
\item{Exported}{\code{\link{chrom_width}}, \code{\link{chrom_smooth}}, \code{\link{chrom_bline}}, \code{\link{chrom_deriv}}, \code{\link{z_optim}},
\code{\link{chrom_amplim}}, \code{\link{chrom_derlims}}, \code{\link{fastchrom_bline}}}
\item{Unexported}{\code{\link{peakmark}}, \code{\link{peakfind}} and its constituent functions}
}
}
