% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_export.R
\name{chrom_export}
\alias{chrom_export}
\title{Compile and export LCQC data and visualizations}
\usage{
chrom_export(
  input_list,
  expath = getwd(),
  plotpars = "default",
  plot_format = "png"
)
}
\arguments{
\item{input_list}{A \strong{named} \code{list} of data to be exported. Must contain \strong{one or more} of the following elements:
\describe{
\item{pks}{Peak detection data output from \code{\link{chrom_detect}}.}
\item{acctops}{Accurate peak apices and inflection points obtained by running \code{\link{dtprep}} on \code{\link{chrom_detect}} output.}
\item{int}{Traditional peak integration data output from \code{\link{chrom_skim}}.}
\item{icf}{Iterative Curve Fitting and integration data output from \code{\link{chrom_icf}}.}
\item{visc}{Dynamic viscosity data output from \code{\link{chrom_visc}}.}
\item{tp}{Theoretical plates and other metrics output from \code{\link{chrom_tplate}}.}
\item{asym}{Asymmetry metrics and Total Peak Analysis (TPA) data output from \code{\link{chrom_asym}}.}
\item{rf}{Retention factors output from \code{\link{chrom_retf}}.}
\item{sf}{Separation factors output from \code{\link{chrom_sepf}}.}
\item{res}{Peak resolution data output from \code{\link{chrom_res}}.}
\item{cperf}{Additional performance metrics output from \code{\link{chrom_addmets}}.}
}}

\item{expath}{The \code{character} string of the \strong{directory} in which to create a folder and export the results.}

\item{plotpars}{A \strong{named} \code{numeric} vector of plotting parameters. \strong{Only relevant when \code{plot_format = "png"}}.
One or more of \code{"w"} (width, in cm), \code{"h"} (height, in cm), \code{"psize"} (point size), and \code{"res"} (resolution).
When set to \code{"default"}, the set of parameters are \code{c(w = 10, h = 12, psize = 12, res = 300)}.}

\item{plot_format}{A \code{character} specifying which format to export plots in. One of \code{"png"} or \code{"pdf"}.}
}
\value{
A timestamped folder is created inside \code{expath} containing the compiled results in a
\strong{LCQC_Results.csv} file, and .png/.pdf visualizations contained in a separate \strong{Plots} directory.
}
\description{
Compiles all key data and visualizations obtainable via various \pkg{lcqc} functions and exports to disk.
}
\examples{
\dontrun{
#Export to the working directory (visualisations in .PNG)
chrom_export(input_list = list(pks = lcqc:::wf_detpeaks,
acctops = dtprep(lcqc:::wf_detpeaks),
int = lcqc:::wf_ints,
icf = lcqc:::wf_icf,
visc = lcqc:::wf_viscs,
tp = lcqc:::wf_tplate,
asym = lcqc:::wf_asyms,
rf = lcqc:::wf_kfs,
sf = lcqc:::wf_sfs,
res = lcqc:::wf_rs,
cperf = lcqc:::wf_perfmets))
}

}
\seealso{
\code{\link{chrom_detect}}, \code{\link{chrom_skim}}, \code{\link{chrom_icf}},
\code{\link{chrom_tplate}}, \code{\link{chrom_asym}}, \code{\link{chrom_retf}}, \code{\link{chrom_sepf}},
\code{\link{chrom_res}}, \code{\link{chrom_addmets}}, \code{\link{read_shim}}, \code{\link{lcqc_render}}
}
