% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_plot.R
\name{chrom_plot}
\alias{chrom_plot}
\title{Plot chromatogram with or without peak markers}
\usage{
chrom_plot(
  chrom_df,
  ptab = NA,
  chrom_vars = "auto",
  id = "auto",
  apex = NA,
  inf = NA,
  ups = NA,
  bound = NA,
  hlt = NA,
  norm_chrom = TRUE,
  cols = "default",
  ptab_mode = "index",
  zoom = "auto",
  which_peaks = NA,
  lablim = NA,
  amp_line = NA,
  xlab = chrom_vars[1],
  ylab = chrom_vars[2],
  plot_title = "Chromatogram",
  draw = TRUE,
  asprat = 0.71
)
}
\arguments{
\item{chrom_df}{A \code{data.frame} containing time (x-axis) and signal (y-axis) variables, whose colnames are included in \code{chrom_vars}.}

\item{ptab}{The peak table containing peak apex/inflection/upslope/boundary point markers, provided as a \code{data.frame}. Often output from function \code{\link{chrom_detect}} or \code{\link{peakfind}}.}

\item{chrom_vars}{A \code{character} vector of length 2 providing column names of x- and y-axis variables present in \code{chrom_df}.}

\item{id}{Column name of the peak ID (index) included in \code{ptab}. If set to \code{"auto"} (default), peak indices are assigned automatically by \code{seq(1,nrow(ptab),1)}.}

\item{apex}{Controls plotting of peak \strong{apices}.
Either a \code{character} vector of length 2 with column names (present in \code{ptab}) containing indices of left and right peak boundaries, or \code{NA} (default).}

\item{inf}{Controls plotting of peak \strong{inflection} points.
Either a \code{character} vector of length 2 with column names (present in \code{ptab}) containing indices of left and right peak boundaries, or \code{NA} (default).}

\item{ups}{Controls plotting of peak \strong{upslope} points.
Either a \code{character} vector of length 2 with column names (present in \code{ptab}) containing indices of left and right upslope points, or \code{NA} (default).}

\item{bound}{Controls plotting of peak \strong{start/end boundaries}.
Either a \code{character} vector of length 2 with column names (present in \code{ptab}) containing indices of left and right peak boundaries, or \code{NA} (default).}

\item{hlt}{A vector of peak apex \strong{indices} present in \code{ptab} that must be highlighted.}

\item{norm_chrom}{A \code{logical} indicating whether the chromatogram signal (y-axis) should be normalized to a maximum value of 100. Defaults to \code{TRUE}.}

\item{cols}{Either \code{"default"} or a \strong{named} vector of recognized colors. Accepted names are: \code{"id"} (peak number), \code{"main"} (main plot),
\code{"apex"} (peak apices), \code{"hlt"} (highlight for apices set in \code{hlt}), \code{"inf"} (inflection points),
\code{"ups"} (upslope points), \code{"bound"} (peak start/end boundaries), \code{"ampline"} (signal amplitude threshold for peak detection).}

\item{ptab_mode}{A \code{character} value denoting whether peak marker data in \code{ptab} is provided as \strong{data indices} (\code{"index"}, default) or \strong{real values}, e.g. retention time (\code{"real"}).}

\item{zoom}{Either a vector of length 2 providing \code{c(low,high)} x-axis limits, or set to \code{"auto"} to automatically zoom into the detected peak region.}

\item{which_peaks}{Either \code{NA} (default) or a \code{numeric} vector of \code{ptab} peak indices to plot, discarding the rest.}

\item{lablim}{Either \code{NA} (default) or a single \code{numeric} value between 0 and 100. Denotes the \strong{relative} signal threshold below which peak apices are not labeled with numbers.}

\item{amp_line}{Amplitude limit to be plotted as a dashed line, given as a single \code{numeric} value. Defaults to \code{NA} (no line plotted).}

\item{xlab}{The x-axis \code{character} label.}

\item{ylab}{The y-axis \code{character} label.}

\item{plot_title}{The main title of the plot (\code{character}).}

\item{draw}{Should the generated plot be shown in the graphics window? \code{TRUE} by default.}

\item{asprat}{Aspect ratio of the plot (defaults to 0.71).}
}
\value{
A \code{ggplot}-class object containing the plot.
}
\description{
Plots a chromatogram with various possible levels of complexity, from a simple plot of the signal versus retention time to inclusion of
peak markers (e.g. peak boundaries, inflection, upslope, and apex points) and the peak detection signal amplitude limit.
}
\examples{
#Simple plot
chrom_plot(lcqc::simlc1)

#Plot with peak markers
\dontrun{
dt <- lcqc:::wf_detpeaks[["results"]]
chrom <- dt[["Chromatogram"]] #Get chromatogram data
pt <- dt[["Peak_Extents"]] #Get peak table
alim <- dt[["Amplitude_Limit"]] #Get amplitude limit
chrom_plot(chrom_df = chrom, ptab = pt, norm_chrom = TRUE, id = "peak", apex = "ind_finmax",
inf = c("ind_linf", "ind_rinf"), ups = c("ind_lups", "ind_rups"),
bound = c("ind_starts", "ind_ends"), amp_line = alim)
}

}
