% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_res.R
\name{chrom_res}
\alias{chrom_res}
\title{Calculate HPLC chromatographic resolution using various methods}
\usage{
chrom_res(
  input,
  peaks1 = "all",
  peaks2 = "all",
  method = "all",
  ks = c(1, "peak"),
  crit_w = "auto",
  verbose_res = FALSE
)
}
\arguments{
\item{input}{The output \code{list} from \code{\link{chrom_detect}}.}

\item{peaks1, peaks2}{Either \code{"all"} (default) or \code{numeric} vectors of \strong{start} and \strong{end} peaks for which to calculate
separation factors. Each element of \code{peaks1} therefore corresponds to the same element of \code{peaks2}. When set to \code{"all"},
resolution is calculated for pairs of successive peaks (e.g. 1:2, 2:3 etc.).}

\item{method}{A \code{character} vector specifying the method(s) to use for the calculations. One or more of: \code{"all"} (default),
\code{"W0"} (full width), \code{"W50_1"} (50\% width, variant 1), \code{"W50_2"} (50\% width, variant 2),
and/or \code{"sepret"} (separation factor and theoretical plate method).}

\item{ks}{Used to configure retention factor calculations. Either a \code{numeric} value of the dead time (in minutes), or a \code{vector} of
length 2 specifying that dead time is to be obtained from the retention time of a specific \code{"peak"}, and providing the \code{numeric}
index of said peak that must be present in \code{input} data.}

\item{crit_w}{The critical width parameter to use for baseline calculation via FastChrom (\code{\link{fastchrom_bline}}).
Defaults to \code{"auto"} or can be set manually as a \code{numeric} (usually equal to the minimum peak width at half height).}

\item{verbose_res}{A \code{logical} switch specifying whether to include intermediate results such as retention factors and
separation factors into the results. Defaults to \code{FALSE}.}
}
\value{
A named \code{list} of length 4 containing a \code{data.frame} of \code{results}, which includes the peak IDs
(\code{id1} and \code{id2}), peak types (\code{type1} and \code{type2}), and resolution values calculated via one or
more methods specified in \code{method} and prefixed with \code{res_}. If \code{verbose_res} is \code{TRUE}, additional
data from \code{\link{chrom_retf}} and \code{\link{chrom_sepf}} are included.
Also included is the dead time \code{t0}, a \code{character} string of various \code{information} about the results,
and the function \code{call} as separate list elements.
}
\description{
Calculates chromatographic resolution using the full width, width at 50\%, and separation factor methods (see \strong{Details}).
}
\details{
Chromatographic resolution \eqn{R} may be calculated between two neighbouring peaks using various approaches which incorporate, in one
form or another, information about both peak apices and widths. Only the peak apices are separated at \eqn{R = 1}, while near-complete
resolution of similarly-sized peaks is achieved at \eqn{R >= 1.5}. For peaks \eqn{1} and a \strong{later-eluting} \eqn{2}, the simpler
and more widespread calculation methods involve the use of retention times \eqn{t_{R1}} and \eqn{t_{R2}}, and corresponding peak widths
at either the base (\eqn{W_1} and \eqn{W_2}) or 50\% peak height (\eqn{W_{0.5h1}} and \eqn{W_{0.5h2}}). Three different equations commonly
utilized with these parameters are presented below. The first of these uses the full peak width at the base and is therefore the most
conservative measure that is relatively more affected by non-Gaussian peak shapes, often resulting in lower values than those obtained
by other methods.
\deqn{R = (t_{R2}-t_{R1})/(0.5\times(W_1+W_2))}
\deqn{R = 1.176\times[(t_{R2}-t_{R1})/(W_{0.5h1}+W_{0.5h2})]}
\deqn{R = (t_{R2}-t_{R1})/[1.7\times 0.5\times(W_{0.5h1}+W_{0.5h2})]}
\strong{For isocratic separations}, where the number of theoretical plates between adjacent peaks should largely be the same, the value
of \eqn{R} may also be related to peak retention factors (\eqn{k_1} and \eqn{k_2}), separation factor \eqn{\alpha}, and average number of
theoretical plates \eqn{\overline{N}} by the following equation (sometimes referred to as the fundamental resolution equation):
\deqn{R = (\sqrt{\overline{N}}/4)\times[(\alpha-1)/\alpha]\times[k_2/(1+k_2)]}
}
\examples{
\dontrun{
#Get data and run
dt <- lcqc:::wf_detpeaks
res <- chrom_res(input = dt, ks = c(1,"peak"))
}

}
\references{
Meyer, V.R. (2010), \emph{Practical High-Performance Liquid Chromatography}, John Wiley & Sons, Chichester, United Kingdom.
}
\seealso{
\code{\link{chrom_sepf}}, \code{\link{chrom_retf}}, \code{\link{chrom_detect}}, \code{\link{fastchrom_bline}}
}
