% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_retf.R
\name{chrom_retf}
\alias{chrom_retf}
\title{Calculation of Retention Factors for HPLC analytes}
\usage{
chrom_retf(input, t0_mode = "peak", t0, peaks = "all", crit_w = "auto")
}
\arguments{
\item{input}{The output \code{list} from \code{\link{chrom_detect}}.}

\item{t0_mode}{The mode to use for calculation of dead time. One of: \code{"peak"} or \code{"manual"}, where \emph{t0} is taken
directly as the time of specified peak maximum or given manually in minutes, respectively.}

\item{t0}{Dead time of the column, i.e. breakthrough time. Either given in \strong{minutes} if \code{t0_mode} is \code{"manual"},
or given as a single peak index included in \code{which_peaks}.}

\item{peaks}{Either \code{"all"} (default) or a \code{numeric} vector of peak indices to calculate retention factors for.}

\item{crit_w}{The critical width parameter to use for baseline calculation via FastChrom (\code{\link{fastchrom_bline}}).
Defaults to \code{"auto"} or can be set manually as a \code{numeric} (usually equal to the minimum peak width at half height).}
}
\value{
A named \code{list} of length 4 containing a \code{data.frame} of \code{results}, which includes the peak \code{id},
peak \code{type} (see \code{\link{chrom_detect}} for descriptions of available peak types), retention time \code{rt}, the
dead time \code{t0}, and the calculated retention factors \code{k}. \strong{If} the input data (from \code{\link{chrom_detect}})
also includes the \code{Compound} names for each peak, these are included as the first column.
Also included is the dead time \code{t0}, a \code{character} string of various \code{information} about the results,
and the function \code{call} as separate list elements.
}
\description{
Calculates retention factors for chromatographic peaks based on column dead time.
See \strong{Details} for further information.
}
\details{
The retention factor \eqn{k} is used to gauge the relative retention of an analyte compared to that of an unretained solute (such as uracil
for reversed-phase separations). It is calculated from dead time \eqn{t_0}, a.k.a. breakthrough time, using the following simple equation:
\eqn{k = (t_R-t_0)/t_0}
Typical RP separations include compounds with \eqn{k} ranging between 1 and 10 (up to 20 for difficult separations). In addition to providing
a retention time manually, it is possible to set the dead time as the retention time of one of the peaks in \pkg{input} when \code{t0_mode} is
set to \code{"peak"}.
}
\examples{
\dontrun{
#Get data and run
dt <- lcqc:::wf_detpeaks
res <- chrom_retf(input = dt, t0_mode = "peak", t0 = 1)
}

}
\references{
Meyer, V.R. (2010), \emph{Practical High-Performance Liquid Chromatography}, John Wiley & Sons, Chichester, United Kingdom.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{fastchrom_bline}}
}
