% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_sepf.R
\name{chrom_sepf}
\alias{chrom_sepf}
\title{Calculation of Separation Factors for HPLC analytes}
\usage{
chrom_sepf(
  input,
  peaks1 = "all",
  peaks2 = "all",
  ks = c(1, "peak"),
  crit_w = "auto"
)
}
\arguments{
\item{input}{The output \code{list} from \code{\link{chrom_detect}}.}

\item{peaks1, peaks2}{Either \code{"all"} (default) or \code{numeric} vectors of \strong{start} and \strong{end} peaks for which to calculate
separation factors. Each element of \code{peaks1} therefore corresponds to the same element of \code{peaks2}. When set to \code{"all"},
separation factors are calculated for pairs of successive peaks (e.g. 1:2, 2:3 etc.).}

\item{ks}{Used to configure retention factor calculations via \code{\link{chrom_retf}}. Usually a \code{vector} of length 2 specifying
the \code{character} mode of calculating retention factors (one of \code{"peak"} or \code{"manual"}) and an accompanying
\code{numeric} value as a parameter. When in \code{"peak"} mode, the accompanying value is the index of a peak
present in \code{input} data whose retention time is used as dead time \emph{t0} to calculate retention factors from.
When mode is \code{"manual"}, the accompanying value must simply be the dead time itself (in minutes). Alternatively,
retention indices may be provided as a \code{numeric} vector of length equal to that of \code{peaks1} and \code{peaks2},
in which case \code{\link{chrom_retf}} is not called.}

\item{crit_w}{The critical width parameter to use for baseline calculation via FastChrom (\code{\link{fastchrom_bline}}).
Defaults to \code{"auto"} or can be set manually as a \code{numeric} (usually equal to the minimum peak width at half height).}
}
\value{
A named \code{list} of length 4 containing a \code{data.frame} of \code{results}, which includes the peak IDs
(\code{id1} and \code{id2}), peak types (\code{type1} and \code{type2}), retention factors \code{k1} and \code{k2},
and the calculated separation factors \code{sep_factor}. \strong{If} the input data (from \code{\link{chrom_detect}})
also includes the \code{Compound} names for each peak pair, these are included as the first two columns.
Also included is the dead time \code{t0}, a \code{character} string of various \code{information} about the results,
and the function \code{call} as separate list elements.
}
\description{
Calculates the retention and separation factors for any number of peak pairs.
}
\details{
Separation factor \eqn{\alpha} (or \emph{relative retention} in some older texts) is useful to determine the relative
degree of separation between any two peaks in a chromatogram. It is therefore a measure of selectivity. For two peaks
\eqn{i} and a \strong{later-eluting} \eqn{j}, \eqn{\alpha} can be calculated as a simple ratio of retention factors
\eqn{k_i} and \eqn{k_j}. If \eqn{\alpha =< 1}, no separation between the components takes place.
\deqn{\alpha = k_j/k_i}
}
\examples{
\dontrun{
#Get data and run
dt <- lcqc:::wf_detpeaks
res <- chrom_sepf(input = dt, peaks1 = c(2:6), peaks2 = c(3:7), ks = c(1,"peak"))
}

}
\references{
Meyer, V.R. (2010), \emph{Practical High-Performance Liquid Chromatography}, John Wiley & Sons, Chichester, United Kingdom.
}
\seealso{
\code{\link{chrom_retf}}, \code{\link{chrom_detect}}, \code{\link{fastchrom_bline}}
}
