% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_skim.R
\name{chrom_skim}
\alias{chrom_skim}
\title{Traditional peak integration of chromatographic data}
\usage{
chrom_skim(
  input,
  method = "gskim",
  skim = 10,
  dyson = 10,
  crit_w = "auto",
  asprat = 0.71,
  plotset = "make"
)
}
\arguments{
\item{input}{The output object of function \code{\link{chrom_detect}}.}

\item{method}{The preferred integration/skimming method. One of: perpendicular drop (\code{"pdrop"}), tangent skim (\code{"tskim"}),
exponential skim (\code{"exskim"}), or Gaussian skim (\code{"gskim"}).}

\item{skim}{The \code{numeric} Skim Ratio criterion to determine suitability of a parent-child peak pair for skimming.
Defaults to \code{10} (see \strong{Details}).}

\item{dyson}{The \code{numeric} Dyson criterion to determine suitability of a parent-child peak pair for skimming.
Defaults to \code{10} (see \strong{Details}).}

\item{crit_w}{The critical width parameter used to calculate peak group baselines via \code{\link{fastchrom_bline}}.}

\item{asprat}{Aspect ratio of the plot (plotted via \code{\link{integ_plot}}). Defaults to \code{0.71}.}

\item{plotset}{A \code{character} string specifying whether data is visualized/shown via \code{\link{integ_plot}}.
One of \code{"make"} (generates plots without printing; default), \code{"print"} (generates and prints plots), or \code{"none"}.}
}
\value{
A \code{list} containing various data in the following named elements:
\describe{
\item{orig_data}{A \code{data.frame} containing the original data indices (\code{"ind"}), retention times (\code{"x"}),
as well as the raw and baseline-corrected signal (\code{"orig_y"} and \code{"y"}, respectively).}
\item{indiv_bln}{A \code{data.frame} containing perpendicular drop and/or skimmed baselines for each individual peak.
The data includes indices (\code{"ind"}), retention time (\code{"x"}), baseline-corrected signal (\code{"y"}),
peak group ID (\code{"group"}), those of parent and child peaks (\code{"peak1"} and \code{"peak2"}),
integration type (\code{"type"}), and a composite ID unique for each baseline (\code{"combo_id"}).}
\item{grp_bln}{A \code{data.frame} containing peak group baselines (one baseline per group). Data include
group indices (\code{"group"}), data indices (\code{"ind"}), retention times (\code{"x"}),
original and baseline-corrected signals (\code{"orig_y"} and \code{"y"}), baselines (\code{"bline"}),
and second derivatives (\code{"sd"}).}
\item{integ_res}{The \strong{main results} \code{data.frame} containing, among other general data such as retention time and signal,
the integrated peak areas (\code{"pa"}) and results of testing for skim suitability via 7 conditions outlined in \strong{Details}
(\code{"true_conds"} and \code{"failed_conds"}).}
\item{max_marks}{A \code{data.frame} containing accurate peak maxima retention time and signals (see \code{\link{acc_max}}) among
other general data such as retention time and signal.}
\item{information}{A \code{character} string of various statements about baseline calculation and integration results.}
\item{call}{The function call.}
\item{plot}{An \strong{optional} element containing the \code{ggplot} visualisation of results (see \code{\link{integ_plot}}).}
}
}
\description{
Creates baseline-resolved, perpendicular drop, and/or various skimmed baselines (tangent/exponential/Gaussian) and
integrates peaks detected via \code{\link{chrom_detect}} via the Trapezoidal Rule. Results may be optionally visualized.
Also see \strong{Details}.
}
\details{
The workflow assesses peaks based on their resolution. First, baselines for fully-resolved peaks are calculated via the FastChrom
algorithm (see \code{\link{fastchrom_bline}} and Johnsen et al., 2013). Common baselines for groups of \strong{un}resolved peaks
are also derived in the same manner. Boundaries between each pair of unresolved peaks are then either separated by a simple
Perpendicular Drop baseline when \code{method = "pdrop"}. Perpendicular Drop tends to increasingly distort peak areas with
decreasing resolution (see Asher et al., 2009). Alternatively, peaks boundaries are tested for suitability for one of three
"skimmed" baseline approaches - tangent skim (\code{method = "tskim"}), exponential skim (\code{"exskim"}), or Gaussian skim
(\code{"gskim"}) - by assessing a series of conditions (Agilent Technologies, 2016; Water Corporation, 2017):
\enumerate{
\item Is the peak boundary classified as either fused or a shoulder?
\item Is at least one peak within the assessed pair classified as fused or a shoulder?
}
If either of these initial conditions are false, no skim is attempted. Otherwise, additional conditions are evaluated:
\enumerate{
\item Is the earlier-occurring peak the parent peak or a child peak (decided based on the height of peak maxima)?
This determines whether to apply a \strong{front} or a \strong{tail} skim.
\item Is the apex of the child peak lower than the closest (along the x-axis) inflection point of the parent peak?
If this conditions is false, \strong{exponential or Gaussian skimming are not attempted}.
\item Is the Skim-Valley ratio (child peak height over that of the inter-peak boundary/valley) lower than the threshold set in \code{skim}?
\item Is the height ratio of the parent peak to the child peak higher than the criterion set in \code{dyson}?
If either of the conditions dependent on \code{skim} and \code{dyson} are false, \strong{skimming is not carried out}.
\item Is the outer boundary of the child peak higher than the inter-peak boundary? If true, \strong{tangent skimming is not possible}.
}
These conditions also determine the identity of the parent and child peak as well as the appropriate skim type
(\strong{front} or \strong{tail}) for each peak pair. Thus, if a boundary is determined to be suitable based on
the above conditions, construction and optimization of a skimmed baseline of the type specified in \code{method} is attempted.

For \strong{front} tangent skim off the parent peak, straight lines are drawn from the inter-peak boundary to each point between
the beginning and the maximum of the child peak (which occurs earlier in this case). For a \strong{tail} tangent skim, lines are
instead drawn between the inter-peak boundary and each point between the maximum and end of the child peak (which is now the later-occurring peak).
Lines whose y-values are >2\% of the maximum child peak signal at any point are discarded. Among the rest, the line whose end-point
is closest to the outer boundary of the child peak and where <40\% of the values are within 1\% of the corresponding chromatographic signal
value is selected.

For exponential skim, the following equation is used to build an exponential curve extending from the inflection point of the
parent peak closest to the inter-peak boundary towards either the start (for \strong{front} skim) or end (for \strong{tail} skim)
of the child peak.
\deqn{H_{ex} = H_0\times exp^{(-B\times(t_R-t_0))} + A\times t_R + C}
Where \eqn{H_{ex}} is the exponential curve value, \eqn{H_0} is the height (signal) at the inter-peak boundary (e.g. valley),
\eqn{B} is the exponential growth/decay function (negative for \strong{tail} skim), \eqn{A} is the slope of the parent peak baseline,
\eqn{C} is the baseline offset of the parent peak, and \eqn{t_R} along with \eqn{t_0} are retention times at \eqn{H_b} and the
inter-peak boundary, respectively. The initial exponential curve is constructed with values of \eqn{B} and \eqn{C} both set to \code{0}.
The offset constant \eqn{C} is then determined by the different between the result and the signal at the inter-peak boundary,
and constant \eqn{B} of the exponential fit is optimized for minimum Euclidean Distance between the curve and original chromatographic
signal \strong{in the parent peak region} (spanning from the closest inflection point to the inter-peak boundary). Finally, the
final exponential curve spanning from the inter-peak boundary to the outer boundary of the child peak is plotted using optimized
constants \eqn{B} and \eqn{C}. This is used as the skimmed baseline.

For Gaussian skim, the following general form of the Gaussian curve is used to construct a model between the apex of the \strong{parent}
peak and the inter-peak boundary.
\deqn{H_{gs} = H_p\times exp^{-(\frac{t_R-t_0}{\sigma})^2}}
Where \eqn{H_{gs}} is the Gaussian curve value, \eqn{H_p} is the parent peak maximum signal, \eqn{t_0} is the corresponding retention
time, \eqn{t_R} is the retention time at \eqn{H_{gs}}, and \eqn{\sigma} is the standard deviation of the Gaussian curve (estimated here
as the half-width of the peak at inflection points). The Gaussian curve is iteratively optimized until Euclidean Distance between the
resulting curve and original parent peak is minimized. The final curve is then constructed using optimized parameters between the
\strong{parent} peak apex and the outer boundary of the \strong{child} peak. The model is then checked for two conditions:
\enumerate{
\item Is the lowest point of the Gaussian curve higher than 1\% of the parent peak maximum?
\item Are any points of the curve in the \strong{child} peak region above the original signal (i.e. does the curve cross the chromatogram)?
}
If either of the above is true, the Gaussian curve is rejected, no further skim is attempted, and a Perpendicular Dropline is instead
constructed. Otherwise, the Gaussian curve is truncated from the \strong{parent} peak maximum until the point where the curve is
of \strong{consistently} lower signal (i.e. height) than the original chromatogram.

Once all the baselines are constructed, the Trapezoidal Rule is used to integrate all peaks and calculate their peak areas (\eqn{PA}).
\deqn{PA = \sum{(x_{i+1}-x_i)\times(y_{i+1}+y_i)/2}}
Results are optionally visualized via \code{\link{integ_plot}}.
}
\examples{
\dontrun{
#Get data and integrate via Perpendicular Drop
det_peaks <- lcqc:::wf_detpeaks
chrom_skim(det_peaks, method = "pdrop")

#Forced exponential skim
chrom_skim(det_peaks, method = "exskim", dyson = 1.5, skim = NA)
}

}
\references{
Agilent Technologies (2016), 'Agilent OpenLAB CDS Data Analysis Reference Guide', document M8410-90031, available at: \url{https://www.agilent.com/cs/library/usermanuals/public/DataAnalysisReferenceGuide.pdf}.

Asher, B.J., D'Agostino, L.A., Way, J.D., Wong, C.S., Harynuk, J.J. (2009), 'Comparison of peak integration methods for the determination of enantiomeric fraction in environmental samples', \emph{Chemosphere} \strong{75} (8), pp. 1042-1048, DOI: \url{https://doi.org/10.1016/j.chemosphere.2009.01.041}.

Dyson, N. (1998), \emph{Chromatographic Integration Methods}, The Royal Society of Chemistry (RSC Chromatography Monographs series), Herefordshire, United Kingdom.

Johnsen, L.G., Skov, T., Houlberg, U., Bro, R. (2013), 'An automated method for baseline correction, peak finding and peak grouping in chromatographic data', \emph{Analyst} \strong{138} (12), pp. 3502-3511, DOI: \url{https://www.doi.org/10.1039/C3AN36276K}.

Kalambet, Y., Kozmin, Y., Samokhin, A. (2018), 'Comparison of integration rules in the case of very narrow peaks', \emph{Chemometrics and Intelligent Laboratory Systems} \strong{179}, pp. 22-30, DOI: \url{https://doi.org/10.1016/j.chemolab.2018.06.001}.

SERAS (2000), 'Standard Operating Procedure 1001: Chromatographic Peak Integration Procedures', U.S. EPA Contract EP-W-09-031, available at: \url{https://clu-in.org/download/ert/1001-r00.pdf} (accessed 22.04.2024).

Waters Corporation (2017), 'Empower Software Data Acquisition and Processing Theory Guide', document 715005481 (Rev. A), available at: \url{https://support.waters.com/KB_Inf/Empower_Breeze/WKB57375_Empower_3_-_How_to_acquire_and_process_data (accessed 19.04.2024)}.
}
\seealso{
This workflow uses a multitude of exported and \strong{un}exported functions:
\describe{
\item{Exported}{\code{\link{dtprep}}, \code{\link{fastchrom_bline}}, \code{\link{integ}}, \code{\link{integ_plot}}}
\item{Unexported}{\code{\link{chkdt}}, skimming functions (\code{\link{tskim}}, \code{\link{exskim}}, \code{\link{gskim}}) and their helper functions, \code{\link{pdrop}}, \code{\link{skimcomp}}}
}
}
