% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_smooth.R
\name{chrom_smooth}
\alias{chrom_smooth}
\title{Chromatogram smoothing}
\usage{
chrom_smooth(x, method = "rect", pts = 3, passes = 3)
}
\arguments{
\item{x}{Input numeric vector of equi-spaced values.}

\item{method}{Smoothing method. One of: \code{"rect"} (rectangular), \code{"tri"} (triangular), \code{"sg_quad"} (S-G quadratic), or \code{"sg_quart"} (S-G quartic, see \strong{Details}).}

\item{pts}{Number of points to use for smoothing (defaults to \code{3}).}

\item{passes}{Number of smoothing passes to apply (defaults to \code{3}).}
}
\value{
The smoothed signal (as a numeric vector).
}
\description{
Smoothes an equi-spaced signal using various common approaches, including rectangular, triangular, and Savitsky-Golay (S-G) smoothing.
}
\details{
The function includes several smoothing approaches. The simplest \code{method}, \code{"rect"}, is a \emph{rectangular boxcar} algoithm,
which replaces each points with the \code{\link[base]{mean}} of several neighboring points. For example, for a 3-point smooth:
\deqn{S_j=\frac{Y_{j-1}+Y_j+Y_{j+1}}{3}}
Triangular smoothing (\code{"tri"}) implements a weighted algorithm. For example, a 5-point smooth becomes:
\deqn{S_j=\frac{Y_{j-2}+2Y_{j-1}+3Y_j+2Y_{j+1}+Y_{j+2}}{9}}
Finally, the Savitsky-Golay (S-G) smoothing methods, also known as digital smoothing polynomial filters, implement moving average-based
smoothing via a quadratic (\code{"sg_quad"}) or quartic (\code{"sg_quart"}) polynomial regression. The latter usually performs better
with narrower peaks, but may distort wider peaks.
}
\examples{
chrom_smooth(lcqc::simlc5[,"Signal"], method = "tri")

}
\references{
O'Haver, T. (2023), \emph{A Pragmatic Introduction to Signal Processing with Applications in Scientific Measurement}, independently published, Maryland, USA, available at: \url{https://terpconnect.umd.edu/~toh/spectrum/} (last accessed 16.04.2024).

Isnanto, R. Rizal (2011), 'Comparison of Several Smoothing Methods in Nonparametric Regression', \emph{Jurnal Sistem Komputer} \strong{1} (1), pp. 41-47.

Savitsky, A., Golay, J.E. (1964), 'Smoothing and Differentiation of Data by Simplified Least Squares Procedures', \emph{Analytical Chemistry} \strong{36} (8), pp. 1627-1639, DOI: \url{https://doi.org/10.1021/ac60214a047}.
}
\seealso{
\code{\link{chrom_width}}
}
