% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_asym.R
\name{chrom_tpa}
\alias{chrom_tpa}
\title{Carry out Total Peak Analysis (TPA) for a selected chromatographic peak}
\usage{
chrom_tpa(
  input,
  vars = c("x", "y"),
  width = "auto",
  accmax = "auto",
  thres = 0.85,
  pknum = "",
  opt = "nlp",
  asprat = 0.71,
  print_plot = TRUE,
  plot_info = FALSE
)
}
\arguments{
\item{input}{A \code{data.frame} containing retention time and signal data of the peak for which TPA is to be carried out.}

\item{vars}{A \code{character} vector of length 2 containing the column names of retention time and signal, in that order.}

\item{width}{Either \code{"auto"} (default) or a single \code{numeric} value of the peak width (in retention time units).}

\item{accmax}{Either \code{"auto"} (default) or a \code{numeric} vector of length 2 containing the retention time and signal
(in that order) of the peak for which TPA is to be carried out.}

\item{thres}{The \code{numeric} threshold between 0 and 0.99 to use for calculation of peak width and its relation to
Gaussian standard deviation for TPA. Defaults to \code{0.85}.}

\item{pknum}{A \code{character} string to append to the plot title. Usually contains information like the peak number and/or compound name.}

\item{opt}{A \code{character} string specifying which method to use for iterative adjustment of the Gaussian model.
One of \code{"optim"} or \code{"nlp"} (default).}

\item{asprat}{Aspect ratio of the plot (defaults to \code{0.71}).}

\item{print_plot}{A \code{logical} switch. Should the generated plot be printed? \code{TRUE} by default.}

\item{plot_info}{A \code{logical} switch. Should extra information about the suitability of the peak for TPA be included in the plot?
\code{FALSE} by default.}
}
\value{
A named \code{list} containing the following elements:
\describe{
\item{results}{A named \code{numeric} vector of key results including the peak width at \code{thres}, the Gaussian standard
deviation (\code{"sigma"}), the \strong{absolute} residual sum (\code{"resid_sum"}), its front- and back-of-peak components
(\code{"resid_front"} and \code{"resid_back"}), corresponding percentage contributions to fronting and tailing
(\code{"percent_fronting"} and \code{"percent_tailing"}), and a binary measure of the suitability of the peak
for TPA (\code{"tpa_suitability"}) determined from the residuals. See \code{\link{chrom_asym}}.}
\item{Peak_Data}{A \code{data.frame} containing the retention time (\code{"rt"}), normalized peak signal (\code{"normpeak"}),
its Gaussian model (\code{"gausspeak"}), residuals between the normalized peak and the Gaussian model (\code{"resids"}),
and their absolute counterparts (\code{"aresids"}).}
\item{Plot}{A \code{ggplot} object containing TPA visualization.}
}
}
\description{
Calculates the Total Peak Analysis (TPA) metrics as described by Wahab et al. (2017) and optionally visualizes the results.
See \code{\link{chrom_asym}} for further information.
}
\examples{
\dontrun{
#Get data
dt <- lcqc:::wf_detpeaks[["results"]][["Peaks"]][["peak_7"]]
chrom_tpa(dt, vars = c("Time","Signal"), pknum = "7 (Placeholder)", plot_info = TRUE)
}

}
\seealso{
\code{\link{chrom_asym}}, \code{\link{peak_hw}}
}
