% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_visc.R
\name{browse_visc}
\alias{browse_visc}
\alias{chrom_visc}
\title{Estimate dynamic viscosity of common HPLC mobile phases}
\usage{
browse_visc()

chrom_visc(ids, fracs, frac_type = "vol", temp = 25)
}
\arguments{
\item{ids}{A \code{character} vector of mobile phase component IDs to calculate viscosity for. One or more of:
butanol (\code{"buoh"}), isopropanol (\code{"ipa"}), acetone (\code{"acet"}), acetonitrile (\code{"mecn"}), benzene (\code{"benz"}),
chloroform (\code{"chcl3"}), cyclohexane (\code{"chex"}), diethyl ether (\code{"dee"}), ethanol (\code{"etoh"}), ethyl acetate (\code{"etac"}),
methanol (\code{"meoh"}), tetrahydrofuran (\code{"thf"}), and water (\code{"h2o"}). If the latter is included, \code{ids} must be of length 2
with the only other possible components being acetonitrile (\code{"mecn"}) or methanol (\code{"meoh"}). See \strong{Details}.}

\item{fracs}{A \code{numeric} vector of mobile phase component fractions of equal length to \code{ids}.}

\item{frac_type}{The \strong{type} of fraction given in \code{fracs}. One of: volumetric (\code{"vol"}, default), mass (\code{"mass"}),
or mole (\code{"mol"}).}

\item{temp}{Temperature to calculate viscosity at. See \strong{Details} and \code{\link{browse_visc}} for acceptable temperature ranges for each solvent.}
}
\value{
A named \code{list} containing \code{results}, a \code{character} string of various \code{information} about them,
and the function \code{call}. The \code{results} element is a \code{list} named according to the mobile phase components
specified in argument \code{ids} and containing a named \code{numeric} vector with the viscosity value (\code{"visc_mPas"}),
temperature taken from \code{temp} (\code{"temp_degC"}), as well as mole, mass, and volume fractions of each component
(suffixed \code{_molfrac}, \code{_massfrac}, and \code{_volfrac}, respectively).
}
\description{
This function estimates the temperature-dependent dynamic viscosity \eqn{\eta} (in mPas) for pure
organic solvents and their mixtures commonly used as mobile phases for High-Performance Liquid Chromatography (HPLC).
It is also possible to estimate viscosities of aqueous mobile phases containing methanol (MeOH) or acetonitrile (MeCN) based on
bilinear interpolation of experimental data collected from various sources. See \strong{Details} for further information.
}
\details{
Dynamic viscosity and density of various organic solvents is calculated using data taken from the Dortmund Data Bank (DDB, 2024)
and packaged with \pkg{lcqc}. Specifically, dynamic viscosity \eqn{\eta} (mPas) is calculated using the Vogel Equation
(García-Colín et al., 1989), which uses experimentally derived empirical coefficients \eqn{A_V}, \eqn{B_V}, and \eqn{C_V}.
The result is also dependent on temperature, given as \eqn{T_K} in Kelvin.
A full list of coefficients and the temperature ranges (\strong{in °C}) within which they are valid is available upon
calling \code{\link{browse_visc}}.
\deqn{\eta\text{ }(mPas) = exp[A_V+(B_V/(C_V + T_K))]}
Similarly, density \eqn{\rho} (kg m^-3) is calculated via the DIPPR105 equation, which incorporates 4 empirical coefficients
(\eqn{A}, \eqn{B}, \eqn{C}, and \eqn{D}) alongside temperature (DDB, 2024; Silva et al., 2018).
\deqn{\rho\text{ }(kg\text{ }m^{-3}) = A/B^{1+(1-T_K/C)^D}}

When a mixture of components is provided in \code{ids}, they have to be accompanied by mole (\code{"mol"}), mass (\code{"mass"}),
or volumetric (\code{"vol"}, default) fractions given in \code{fracs}. Whichever fraction type (\code{frac_type}) is chosen,
the others are calculated and included in function output. For example, conversion to \eqn{F_{mass}} and \eqn{F_{mol}} from
\eqn{F_{vol}} is carried out using density \eqn{\rho} and relative molecular mass (\eqn{RMM}, g/mol) as follows for each component.
\deqn{F_{mass} = F_{vol}\times\rho}
\deqn{F_{mol} = F_{mass}/RMM}
The overall viscosity \eqn{\eta_{total}} of the mixtures is calculated using the Linear Blend Rule \strong{for purely organic mixtures}.
Briefly, the contributing \strong{mole} fraction (\eqn{F_{mol}}) of each component \eqn{i} is multiplied by the corresponding viscosity
value of the pure component, and results for all components obtained in this manner are summed.
\deqn{\eta_{total} = \sum{F_{mol(i)}*\eta_i}}
For \strong{organic-water mixtures}, the Linear Blend Rule is not applicable due to the strong interactions between water molecules
(Snyder et al., 1997). For this reason, \pkg{lcqc} is limited to estimating the viscosity of such mixtures via
\strong{bilinear interpolation} of experimental data obtained for the two most popular aqueous mobile phases in HPLC,
\strong{Methanol-Water} and \strong{Acetonitrile-Water}, from various sources (Snyder et al., 2009; Thompson et al., 2006;
Teutenberg et al., 2009; Wohlfarth & Wohlfahrt, 2001).
}
\examples{
#Pure component (MeOH)
vpure <- chrom_visc(ids = "meoh", fracs = 1, frac_type = "vol", temp = 25)

#Mixture (3:7 MeOH:EtAc)
vmix <- chrom_visc(ids = c("meoh", "etac"), fracs = c(0.3, 0.7), frac_type = "vol", temp = 25)

#Aqueous mobile phase (2:8 MeOH:H2O)
vaq <- chrom_visc(ids = c("meoh", "h2o"), fracs = c(0.2, 0.8), frac_type = "vol", temp = 25)

}
\references{
Dortmund Data Bank (2024), 'Online Calculation', available at: \url{https://www.ddbst.com/calculation.html} (accessed 25.04.2024).

García-Colín, L.S., del Castillo, L.F., Goldstein, P. (1989), 'Theoretical Basis for the Vogel-Fulcher-Tammann Equation', \emph{Physical Review B} \strong{40} (10), pp. 7040-7044, DOI: \url{https://www.doi.org/10.1103/PhysRevB.40.7040}.

Silva, M., Vieira, B., Ottens, M. (2018), 'Preferential crystallization for the purification of similar hydrophobic polyphenols', \emph{Journal of Chemical Technology & Biotechnology} \strong{93} (7), pp. 1997-2010, DOI: \url{https://doi.org/10.1002/jctb.5526}.

Snyder, L.R., Kirkland, J.J., Glajch, J.L. (1997), 'Appendix II: Properties of Solvent Used in HPLC', In: \emph{Practical HPLC Method Development, Second Edition}, John Wiley & Sons, USA.

Snyder, L.R., Kirkland, J.J., Dolan, J.W. (2009), 'Appendix I: Properties of HPLC Solvents', In: \emph{Introduction to Modern Liquid Chromatography}, pp. 879-886, DOI: \url{https://doi.org/10.1002/9780470508183.app1}.

Teutenberg, T., Wiese, S., Wagner, P., Gmehling, J. (2009), 'High-temperature liquid chromatography. Part II: Determination of the viscosities of binary solvent mixtures - Implications for liquid chromatographic separations', \emph{Journal of Chromatography A} \strong{1216} (48), pp. 8470-8479, DOI: \url{https://doi.org/10.1016/j.chroma.2009.09.075}.

Thompson, J.W., Kaiser, T.J., Jorgenson, J.W. (2006), 'Viscosity measurements of methanol-water and acetonitrile-water mixtures at pressures up to 3500 bar using a novel capillary time-of-flight viscometer', \emph{Journal of Chromatography A} \strong{1134} (1), pp. 201-209, DOI: \url{https://doi.org/10.1016/j.chroma.2006.09.006}.

Wohlfarth, C., Wohlfahrt, B. (2001), '3 Mixtures of Water and Organic Compounds', In: \emph{Landolt-Börnstein - Group IV Physical Chemistry, Volume 18A: Pure Organometallic and Organononmetallic Liquids, Binary Liquid Mixtures}, Springer-Verlag Berlin Heidelberg, DOI: \url{https://www.doi.org/10.1007/10639275_6}.
}
\seealso{
\code{\link{chrom_tplate}}
}
