% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_width.R
\name{chrom_width}
\alias{chrom_width}
\title{Auto-estimate smoothing width and number of smoothing passes}
\usage{
chrom_width(
  rtime,
  sig,
  start_smooth = c(7, 1),
  bcorr = "none",
  bpars = "default",
  ampfrac = 1,
  smooth_method = rep("tri", 2)
)
}
\arguments{
\item{rtime}{A \code{numeric} vector of retention time (x-axis) values.}

\item{sig}{A \code{numeric} vector of signal (y-axis) values.}

\item{start_smooth}{A \code{numeric} vector of length 2 containing, in that order, the \strong{odd} number of smoothing points and passes to use for \strong{initial smoothing}. Defaults to \code{c(7,1)}.}

\item{bcorr}{Method to use for baseline correction (defaults to \code{"none"}), see \code{\link{chrom_bline}} for available options.}

\item{bpars}{Parameters to use for baseline correction when \code{bcorr!="none"}.}

\item{ampfrac}{The \code{numeric} quantile percentage value (from 0.001 to 50) to use for amplitude limit estimation via function \code{\link{chrom_amplim}}.}

\item{smooth_method}{A \code{character} vector of length 2 providing smoothing methods for signal and derivatives (in that order).
If a vector of length 1 is given, the same method is applied to both types of data. See \code{\link{chrom_smooth}} for available methods.}
}
\value{
A named \code{numeric} vector of length 5 containing the following elements:
\describe{
\item{"points"}{Number of auto-estimated smoothing points (smooth width).}
\item{"passes"}{Number of auto-estimated smoothing passes.}
\item{"width_5_mean"}{\strong{Average} peak width at 5\% of peak height (in number of points/indices).}
\item{"width_5_min"}{\strong{Minimum} peak width at 5\% of peak height (in number of points/indices).}
\item{"width_inf"}{Minimum inflection point width (in number of points/indices).}
}
}
\description{
Part of the \code{\link{chrom_detect}} workflow. Estimates smoothing window width, number of smoothing passes, average and minimum peak width at 5\% height, minimum width at inflection points.
}
\examples{
dt <- lcqc::simlc1
chrom_width(dt[,"Time"], dt[,"Signal"], ampfrac = 0.05)

}
\seealso{
\code{\link{chrom_detect}}, \code{\link{acc_inf}}
}
