% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{corresp}
\alias{corresp}
\title{Detect peak markers corresponding to another single peak marker}
\usage{
corresp(peak_val, compvec, minsim = 3)
}
\arguments{
\item{peak_val}{A single \code{numeric} index of a peak apex.}

\item{compvec}{A \code{numeric} vector of indices to compare \code{peak_val} to.}

\item{minsim}{The difference in indices between \code{peak_val} and \code{compvec} to test for. Defaults to \code{3}.}
}
\value{
A single \code{numeric} value of a \code{compvec} index matching \code{peak_val}, or \code{NA} if no match is found.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

The function detects a sufficiently similar peak index among \code{compvec} (if any) to a reference value (\code{peak_val}).
The similarity is determined by difference in index given in \code{minsim}.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}
}
\keyword{internal}
