% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cprint}
\alias{cprint}
\title{Print information about function run}
\usage{
cprint(x)
}
\arguments{
\item{x}{Output from one of: \code{chrom_detect}, \code{chrom_skim}, \code{chrom_icf}, \code{chrom_visc},
\code{chrom_tplate}, \code{chrom_asym}, \code{chrom_retf}, \code{chrom_sepf}, \code{chrom_res}, or \code{chrom_addmets}.}
}
\value{
A \code{cat} printout in the console.
}
\description{
Prints the contents of \code{$information} contained in the output of various \pkg{lcqc} functions.
}
\examples{
\dontrun{
cprint(lcqc:::wf_detpeaks)
}
}
