% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{det_infs}
\alias{det_infs}
\title{Detect closest inflection points for one or more peak apices}
\usage{
det_infs(sig, sdmin, cross, upcross, cross_nc = NA, upcross_nc = NA)
}
\arguments{
\item{sig}{Signal \code{numeric} vector.}

\item{sdmin}{Indices of second derivative minima along \code{sig}.}

\item{cross, upcross}{Indices of second derivative downcrosses and upcrosses along \code{sig}.}

\item{cross_nc, upcross_nc}{\strong{Non-compliant} equivalents to \code{cross} and \code{upcross}.}
}
\value{
A named \code{list} containing \code{left} and \code{right} inflection points for each \code{sdmin}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

An extension of \code{\link{raw_infs}} used to detect both left and right inflection points for one or more peaks.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{raw_infs}}, \code{\link{markcross_chk}}
}
\keyword{internal}
