% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{detect_bunch}
\alias{detect_bunch}
\title{Detect and mitigate bunching within second derivative minima}
\usage{
detect_bunch(sdmin, sigmax, max_w = NA)
}
\arguments{
\item{sdmin}{A \code{numeric} vector containing data indices of second derivative minima.}

\item{sigmax}{A \code{numeric} vector containing data indices of signal maxima.}

\item{max_w}{The maximum smooth width (in point indices); 1/4th of this value is used as the similarity threshold for bunching (see \strong{Details}).}
}
\value{
A numeric vector of filtered \code{sdmin} values.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Removes bunched second derivative minima prior to attempting peak detection.
}
\details{
The function detects groups of 2 or more second derivative (SD) minima (\code{sdmin}) which are less than \code{0.25*max_w} apart.
Each SD minimum within these groups is tested for proximal signal maxima. If only one peak has a single proximal maximum,
all others within the group are removed as erroneous. If either none or more than one corresponding \code{sigmax} are present, the peak index closest
to the median of all those within a group is taken as the final value, and all other are removed as erroneous.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}
}
\keyword{internal}
