% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtprep.R
\name{dtprep}
\alias{dtprep}
\title{Summarize and prepare peak detection data for integration}
\usage{
dtprep(input, vars = "auto", crit_w = "auto")
}
\arguments{
\item{input}{The output of function \code{\link{chrom_detect}}.}

\item{vars}{Either \code{"auto"} (default) or a \code{character} vector of length 3 specifying the column names of
retention time, signal, and second derivative data in \code{input} (in that order).}

\item{crit_w}{Critical width parameter to use for baseline construction via \code{\link{fastchrom_bline}}.}
}
\value{
A \code{list} of length 5 containing the following elements:
\describe{
\item{main_df}{A \code{data.frame} containing data indices, retention time, original and baseline-corrected signal (output from \code{\link{fastchrom_bline}}).}
\item{type_df}{A \code{list} of 5 \code{data.frame} objects summarizing types, original, baseline-corrected (\code{\link{fastchrom_bline}}),
and/or accurately-interpolated (\code{\link{acc_inf}}, \code{\link{acc_max}}) values of peak \code{starts}, \code{ends}, \code{linfs}, \code{rinfs}, \code{maxes}.}
\item{grp_df}{Data equivalent to that in \code{type_df} but organized by groups of baseline-resolved peaks in a nested \code{list}.}
\item{grp_blines}{A \code{data.frame} containing \strong{baseline-resolved peak group IDs}, original xy-coordinates, FastChrom baselines,
baseline-corrected signal, and second derivatives of all peak regions.}
\item{acc_tops}{A \code{data.frame} containing combined results of accurate inflection point and peak retention time determination
via \code{\link{acc_inf}} and \code{\link{acc_max}}, respectively.}
\item{peak_list}{A \code{list} of \code{data.frame} objects, each containing retention time, original signal, baselines, baseline-corrected signal,
second derivative, and peak ID data for an individual peak.}
}
}
\description{
This function takes peak detection results output of \code{\link{chrom_detect}} and prepares the data for integration via functions
\code{\link{chrom_skim}} and/or \code{\link{chrom_icf}}.
}
\examples{
\dontrun{
dtprep(lcqc:::wf_detpeaks)
}

}
\seealso{
\code{\link{chrom_skim}}, \code{\link{chrom_icf}}, \code{\link{chrom_detect}}, \code{\link{acc_inf}}, \code{\link{acc_max}}
}
