% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exgc1}
\alias{exgc1}
\title{Experimental GC-MS Chromatogram (example 1)}
\format{
\subsection{\code{exgc1}}{

A data frame with 2141 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exgc1
}
\description{
An experimental example two column GC-MS chromatogram (1 of 2)
}
\keyword{datasets}
