% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exgc2}
\alias{exgc2}
\title{Experimental GC-MS Chromatogram (example 2)}
\format{
\subsection{\code{exgc2}}{

A data frame with 2141 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exgc2
}
\description{
An experimental example two column GC-MS chromatogram (2 of 2)
}
\keyword{datasets}
