% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc1}
\alias{exlc1}
\title{Experimental LC Chromatogram (example 1)}
\format{
\subsection{\code{exlc1}}{

A data frame with 1301 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc1
}
\description{
An experimental example two column HPLC chromatogram (1 of 11)
}
\keyword{datasets}
