% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc10}
\alias{exlc10}
\title{Experimental LC Chromatogram (example 10)}
\format{
\subsection{\code{exlc10}}{

A data frame with 1102 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc10
}
\description{
An experimental example two column HPLC chromatogram (10 of 11)
}
\keyword{datasets}
