% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc11}
\alias{exlc11}
\title{Experimental LC Chromatogram (example 11)}
\format{
\subsection{\code{exlc11}}{

A data frame with 3282 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc11
}
\description{
An experimental example two column HPLC chromatogram (11 of 11)
}
\keyword{datasets}
