% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc2}
\alias{exlc2}
\title{Experimental LC Chromatogram (example 2)}
\format{
\subsection{\code{exlc2}}{

A data frame with 1301 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc2
}
\description{
An experimental example two column HPLC chromatogram (2 of 11)
}
\keyword{datasets}
