% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc3}
\alias{exlc3}
\title{Experimental LC Chromatogram (example 3)}
\format{
\subsection{\code{exlc3}}{

A data frame with 1301 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc3
}
\description{
An experimental example two column HPLC chromatogram (3 of 11)
}
\keyword{datasets}
