% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc4}
\alias{exlc4}
\title{Experimental LC Chromatogram (example 4)}
\format{
\subsection{\code{exlc4}}{

A data frame with 1301 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc4
}
\description{
An experimental example two column HPLC chromatogram (4 of 11)
}
\keyword{datasets}
