% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc5}
\alias{exlc5}
\title{Experimental LC Chromatogram (example 5)}
\format{
\subsection{\code{exlc5}}{

A data frame with 1353 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc5
}
\description{
An experimental example two column HPLC chromatogram (5 of 11)
}
\keyword{datasets}
