% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc6}
\alias{exlc6}
\title{Experimental LC Chromatogram (example 6)}
\format{
\subsection{\code{exlc6}}{

A data frame with 4701 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc6
}
\description{
An experimental example two column HPLC chromatogram (6 of 11)
}
\keyword{datasets}
