% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc7}
\alias{exlc7}
\title{Experimental LC Chromatogram (example 7)}
\format{
\subsection{\code{exlc7}}{

A data frame with 664 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc7
}
\description{
An experimental example two column HPLC chromatogram (7 of 11)
}
\keyword{datasets}
