% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc8}
\alias{exlc8}
\title{Experimental LC Chromatogram (example 8)}
\format{
\subsection{\code{exlc8}}{

A data frame with 1596 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc8
}
\description{
An experimental example two column HPLC chromatogram (8 of 11)
}
\keyword{datasets}
