% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exlc9}
\alias{exlc9}
\title{Experimental LC Chromatogram (example 9)}
\format{
\subsection{\code{exlc9}}{

A data frame with 1436 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
exlc9
}
\description{
An experimental example two column HPLC chromatogram (9 of 11)
}
\keyword{datasets}
