% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{experr}
\alias{experr}
\title{Error minimization function for exponential skim}
\usage{
experr(B_opt, extra_pars, marks, opt_data, opt_method = 1)
}
\arguments{
\item{B_opt}{Exponential growth/decay constant to be optimized. Starting value is set to 0 (no growth or decay).}

\item{extra_pars}{A \strong{named} \code{numeric} vector of length 4 containing values for constant A (slope of the parent peak baseline, \code{"fin_A"}),
constant C (y-offset of the baseline of the parent peak, \code{"fin_C"}), and the x- and y-data at the valley or other boundary between the peaks
(\code{"x0"} and \code{"y0"}).}

\item{marks}{A \strong{named} \code{numeric} vector of length 3 containing data indices of the starting (\code{"infind"}) and ending (\code{"fitend"})
points of the exponential curve, as well as the location of the valley (or other boundary) between the parent and child peaks (\code{"vlind"}).}

\item{opt_data}{A \code{data.frame} containing the output of \code{\link{dtprep}}\code{$grp_blines} for the \strong{parent and child} peaks only.}

\item{opt_method}{One of \code{1} or \code{2}. Determines the type of error measure calculated for exponential curve.
When \code{1}, Euclidean distance between the exponential curve and original data in the \strong{parent peak region} is calculated.
When \code{2}, the difference of exponential curve \strong{endpoint} from zero is calculated.}
}
\value{
The single \code{numeric} error measure.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Calculates two different measures of error for the exponential curve. Intended to use with \code{\link{exskim}} for error minimization via
\code{\link[stats]{optim}}. See \code{\link{chrom_skim}} for further details.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{exskim}}, \code{\link{expfit}}
}
\keyword{internal}
