% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{expfit}
\alias{expfit}
\title{Fit an exponential curve (for exponential skimming)}
\usage{
expfit(B, input, y0, x0, A, C)
}
\arguments{
\item{B}{Growth/decay constant of the exponential curve.}

\item{input}{A \code{numeric} vector containing retention time (x-axis) for which to derive an exponential curve.}

\item{x0, y0}{Retention time (x-axis) and baseline-corrected signal (y-axis) values at the valley (or other boundary) between parent and child peaks.}

\item{A}{Slope of the baseline of the parent peak.}

\item{C}{Offset of the baseline of the parent peak.}
}
\value{
A \code{numeric} vector of the exponential curve equal in length to \code{input}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Constructs an exponential curve. Used for iterative optimization of exponential skim in \code{\link{exskim}}.
See \code{\link{chrom_skim}} for further details.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{experr}}, \code{\link{exskim}}
}
\keyword{internal}
