% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{exskim}
\alias{exskim}
\alias{gskim}
\title{Front or tail Exponential or Gaussian Skim on a pair of peaks}
\usage{
exskim(peaks, maxs, lin, rin, skim_type, h2 = "auto")

gskim(peaks, maxs, lin, rin, skim_type, h2 = "auto")
}
\arguments{
\item{peaks}{A \code{list} of length 2, where each element is a \code{data.frame} containing the data indices (\code{"ind"}), retention times
(\code{"x"}), signal (\code{"y"}), and baseline (\code{"bline"}) values for a single peak (either parent or child).}

\item{maxs}{A 2-row \code{data.frame} containing key data about peak apices as output from \code{\link{dtprep}}\code{$type_df$maxes}.}

\item{lin, rin}{Both 2-row \code{data.frame} objects containing key data about peak inflection points as output from
\code{\link{dtprep}}\code{$type_df$linfs} and \code{\link{dtprep}}\code{$type_df$rinfs}, respectively.}

\item{skim_type}{Type of skim to apply. One of \code{"front"} or \code{"tail"}.}

\item{h2}{A \code{numeric} ID of the child peak. Either \code{1} or \code{2} (earlier or later peak, respectively).
May also be set to \code{"auto"} (default), in which case the peak with the higher signal at the apex is chosen as the parent peak.}
}
\value{
A \code{data.frame} containing original data indices (\code{"ind"}), retention time (\code{"x"}), baseline-corrected signal (\code{"y"}),
peak group ID (\code{"group"}), those of the parent and child peak (\code{"peak1"} and \code{"peak2"}), the skim type (one of \code{"EF"} or
\code{"ET"} for front and tail exponential skims; \code{"GF"} or \code{"GT"} for front and tail Gaussian skims, respectively),
and the combined ID used for plotting via \code{\link{integ_plot}} (\code{"combo_id"}).
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Calculates an Exponential or Gaussian Skim for a suitable parent-child pair of chromatographic peaks.
For further details, see \code{\link{chrom_skim}}.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{experr}}, \code{\link{expfit}}, \code{\link{gskimerr}}, \code{\link{fit_1G}}
}
\keyword{internal}
