% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_lcqc.R
\name{ext_lcqc}
\alias{ext_lcqc}
\title{Browse external package data.}
\usage{
ext_lcqc(full.path = TRUE, txt_only = TRUE)
}
\arguments{
\item{full.path}{A \code{logical} switch. Should full paths to external files be returned? Defaults to \code{TRUE}.}

\item{txt_only}{A \code{logical} switch.. Should only \strong{.txt} files be listed? When \code{FALSE}, all external files included with the package are listed.}
}
\value{
Filenames (or file paths if \code{full.path = TRUE}) of external data files included with the package.
}
\description{
Lists paths or filenames of external raw data files (such as .csv and .txt) included with the \pkg{lcqc} package.
}
\examples{
ext_lcqc(full.path = FALSE)
}
