% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{find_fused}
\alias{find_fused}
\title{Detect fused boundaries from initial ApexTrack expansion results}
\usage{
find_fused(input, rtime, sdmax, sigmin)
}
\arguments{
\item{input}{The output of \code{\link{apexbnds}} containing estimates of baseline-resolved boundaries (preferably originating from \code{\link{base_raw}}).}

\item{rtime}{A \code{numeric} vector of retention time (x-axis) data.}

\item{sdmax}{Indices of second derivative maxima along \code{rtime}.}

\item{sigmin}{Indices of signal minima along \code{rtime}.}
}
\value{
A \code{list} of length 2 containing:
\describe{
\item{bounds}{Named \code{numeric} vector of fused peak indices.}
\item{peak_inds}{Peak numbers corresponding to indices in list element \code{bounds}.}
}
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

This function is intended to be used after \code{base_raw} to detect expanded baseline intersections and find fused peak boundaries
(only those characterised by a valley/signal minimum, not shoulder or round peaks).
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{base_raw}}, \code{\link{apexbnds}}
}
\keyword{internal}
