% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Detect maxima in vector}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{Numeric vector in which to detect maxima.}

\item{m}{The number of data points on either side (i.e. neighbourhood of points) used to validate maxima and minima.}
}
\value{
Position(s) of maxima in vector.
}
\description{
Detects maximum values (peaks) in a series of numbers.
}
\details{
The function may also be used to detect minima by simple passing the inverse of a vector, i.e. \code{-x}.
}
\examples{
vec <- c(1,2,3,2,1,1,2,1)
find_peaks(vec, m = 3)
find_peaks(-vec, m = 3)
}
