% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{find_shld}
\alias{find_shld}
\title{Detect shoulder and round peak boundaries (when all others are available)}
\usage{
find_shld(bnds, infs, sdmin, sdmax, fder, sder, fd_thres, sd_thres)
}
\arguments{
\item{bnds}{\strong{Final} start and end \code{numeric} indices of baseline-resolved peaks.}

\item{infs}{A \code{numeric} vector of all inflection point indices.}

\item{sdmin}{A \code{numeric} vector of second derivative minima (representing peaks).}

\item{sdmax}{A \code{numeric} vector of second derivative maxima indices.}

\item{fder, sder}{Both \code{numeric} vectors of first (\code{fder}) and second (\code{sder}) derivatives.}

\item{fd_thres}{A \code{numeric} vector of first derivative peak detection thresholds (e.g. obtained via \code{\link{chrom_derlims}}).}

\item{sd_thres}{Second derivative thresholds similar to \code{fd_thres}.}
}
\value{
A \code{named} \code{numeric} vector of shoulder and round peak indices, named \code{"S"} and \code{"R"}, respectively.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

The function is designed to be applied after all of \code{\link{base_raw}}, \code{\link{find_fused}}, and \code{\link{base_fused}} and
detects shoulder and round peak boundaries.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{chrom_derlims}}
}
\keyword{internal}
