% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{find_ups}
\alias{find_ups}
\title{Find upslope points for one or more peaks}
\usage{
find_ups(lbds, rbds, linfs, rinfs, sdmin, upsmrk, sder)
}
\arguments{
\item{lbds, rbds}{Indices of peak start (\code{lbds}) and end (\code{rbds}) boundaries.}

\item{linfs, rinfs}{Indices of left and right inflection points. Must each be equal in length to that of \code{sdmin}.}

\item{sdmin}{Indices of second derivative minima (\code{numeric}).}

\item{upsmrk}{Indices of second derivative maxima (\code{numeric}) characteristic of upslope points.}

\item{sder}{Second derivative data of the signal (\code{numeric}).}
}
\value{
A \code{list} with \code{$left} and \code{$right} upslope points.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Finds upslope points characterised by second derivative maxima, provided peak apices, inflection points, and start/end indices (boundaries).
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}
}
\keyword{internal}
