% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{fit_1G}
\alias{fit_1G}
\title{Fit a Gaussian curve (for Gaussian skimming)}
\usage{
fit_1G(apex, loc, sigma, data)
}
\arguments{
\item{apex}{Height (signal at apex) of the peak.}

\item{loc}{Retention time at peak apex.}

\item{sigma}{Constant denoting the standard deviation of the Gaussian distribution.}

\item{data}{A \code{numeric} vector of retention time (x-axis) values for which to construct the model.}
}
\value{
A \code{numeric} vector of the Gaussian curve equal in length to \code{data}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Constructs a Gaussian curve. Used for iterative optimization of Gaussian skim in \code{\link{gskim}}.
See \code{\link{chrom_skim}} for further details.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{gskimerr}}, \code{\link{gskim}}
}
\keyword{internal}
