% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{gskimerr}
\alias{gskimerr}
\title{Error minimization function for exponential skim}
\usage{
gskimerr(pars, input)
}
\arguments{
\item{pars}{A \code{numeric} vector of parameters including, in that order, the retention time at parent peak apex,
the standard deviation of the Gaussian curve, and the signal (y-value) at the parent peak apex.}

\item{input}{A \code{data.frame} containing the output of \code{\link{dtprep}}\code{$grp_blines} for the region spanning
from the maximum of the \strong{parent} peak to the valley (or other boundary) between the parent and child peaks.}
}
\value{
A single \code{numeric} error measure (Euclidean Distance) between the Gaussian model and baseline-corrected chromatogram.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Calculates the Euclidean Distance between the Gaussian model and chromatographic data as error measure for
iterative optimization of the Gaussian skim (\code{\link{gskim}}). For further details, see \code{\link{chrom_skim}}.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{fit_1G}}, \code{\link{gskim}}
}
\keyword{internal}
