% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{icf_EGH}
\alias{icf_EGH}
\title{Build an Exponential-Gaussian Hybrid (EGH) model for one or more peaks}
\usage{
icf_EGH(
  H,
  center,
  sigma,
  tau,
  H2 = NA,
  center2 = NA,
  sigma2 = NA,
  tau2 = NA,
  H3 = NA,
  center3 = NA,
  sigma3 = NA,
  tau3 = NA,
  H4 = NA,
  center4 = NA,
  sigma4 = NA,
  tau4 = NA,
  H5 = NA,
  center5 = NA,
  sigma5 = NA,
  tau5 = NA,
  t
)
}
\arguments{
\item{H, H2, H3, H4, H5}{The height at peak maximum for 1-5 peaks.}

\item{center, center2, center3, center4, center5}{The retention time at peak maximum for 1-5 peaks.}

\item{sigma, sigma2, sigma3, sigma4, sigma5}{The standard deviation of the Gaussian component of the curve for 1-5 peaks.}

\item{tau, tau2, tau3, tau4, tau5}{The time constant of the truncated exponential component of the curve for 1-5 peaks.}

\item{t}{A \code{numeric} vector of peak retention times.}
}
\value{
A \code{list} of \code{numeric} vectors of EGH curve modelled signal corresponding to and equal
in length to retention time \code{t}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_icf}} workflow, not intended for standalone use.

Builds an Exponential-Gaussian Hybrid curve for up to 5 peaks to be used in conjunction with error calculation function \code{\link{egh_penalty}}.
Initial standard deviation and time constant for the curve may be estimated via \code{\link{init_egh}}.
See \code{\link{chrom_icf}} for details.
}
\seealso{
\code{\link{chrom_icf}}, \code{\link{egh_penalty}}, \code{\link{init_egh}}
}
\keyword{internal}
