% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{icf_EMG}
\alias{icf_EMG}
\title{Build an Exponentially-Modified Gaussian (EMG) curve}
\usage{
icf_EMG(
  A,
  center,
  sigma,
  tau,
  A2 = NA,
  center2 = NA,
  sigma2 = NA,
  tau2 = NA,
  A3 = NA,
  center3 = NA,
  sigma3 = NA,
  tau3 = NA,
  A4 = NA,
  center4 = NA,
  sigma4 = NA,
  tau4 = NA,
  A5 = NA,
  center5 = NA,
  sigma5 = NA,
  tau5 = NA,
  t,
  variant = "emg1"
)
}
\arguments{
\item{A, A2, A3, A4, A5}{The peak area for 1-5 peaks.}

\item{center, center2, center3, center4, center5}{The retention time at peak maximum for 1-5 peaks.}

\item{sigma, sigma2, sigma3, sigma4, sigma5}{The standard deviation of the Gaussian component of the curve for 1-5 peaks.}

\item{tau, tau2, tau3, tau4, tau5}{The time constant of the truncated exponential component of the curve for 1-5 peaks.}

\item{t}{A \code{numeric} vector of peak retention times.}

\item{variant}{The chosen expression variant of the EMG function as a \code{character} string (sometimes affects optimization stability).
Current variants are \code{"emg1"} or \code{"emg2"}.}
}
\value{
A \code{list} of \code{numeric} vectors of EMG curve modelled signal corresponding to and equal
in length to retention time \code{t}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_icf}} workflow, not intended for standalone use.

Builds an Exponentially-Modified Gaussian curve for up to 5 peaks to be used in conjunction with error calculation function \code{\link{emg_penalty}}.
See \code{\link{chrom_icf}} for details.
}
\seealso{
\code{\link{chrom_icf}}, \code{\link{emg_penalty}}
}
\keyword{internal}
