% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{icf_ETG}
\alias{icf_ETG}
\title{Build an Empirically-Transformed Gaussian (ETG) curve}
\usage{
icf_ETG(
  H,
  tl,
  tt,
  kl,
  kt,
  laml,
  lamr,
  a,
  b,
  H2 = NA,
  tl2 = NA,
  tt2 = NA,
  kl2 = NA,
  kt2 = NA,
  laml2 = NA,
  lamr2 = NA,
  a2 = NA,
  b2 = NA,
  H3 = NA,
  tl3 = NA,
  tt3 = NA,
  kl3 = NA,
  kt3 = NA,
  laml3 = NA,
  lamr3 = NA,
  a3 = NA,
  b3 = NA,
  H4 = NA,
  tl4 = NA,
  tt4 = NA,
  kl4 = NA,
  kt4 = NA,
  laml4 = NA,
  lamr4 = NA,
  a4 = NA,
  b4 = NA,
  H5 = NA,
  tl5 = NA,
  tt5 = NA,
  kl5 = NA,
  kt5 = NA,
  laml5 = NA,
  lamr5 = NA,
  a5 = NA,
  b5 = NA,
  t
)
}
\arguments{
\item{H, H2, H3, H4, H5}{The peak area for 1-5 peaks.}

\item{tl, tl2, tl3, tl4, tl5}{The retention times of \strong{left} inflection points for 1-5 peaks. \strong{This parameter remains constant}.}

\item{tt, tt2, tt3, tt4, tt5}{The retention times of \strong{right} inflection points for 1-5 peaks. \strong{This parameter remains constant}.}

\item{kl, kl2, kl3, kl4, kl5}{Parameter of the \strong{leading} edge for 1-5 peaks, originally defined by a reciprocal of Gaussian standard deviation.}

\item{kt, kt2, kt3, kt4, kt5}{Parameter of the \strong{trailing} edge for 1-5 peaks, originally defined by a reciprocal of Gaussian standard deviation.}

\item{laml, laml2, laml3, laml4, laml5}{Parameter two of the \strong{leading} edge for 1-5 peaks.}

\item{lamr, lamr2, lamr3, lamr4, lamr5}{Parameter two of the \strong{trailing} edge for 1-5 peaks.}

\item{a, a2, a3, a4, a5}{Parameter three of the \strong{leading} edge for 1-5 peaks.}

\item{b, b2, b3, b4, b5}{Parameter three of the \strong{trailing} edge for 1-5 peaks.}

\item{t}{A \code{numeric} vector of peak retention times.}
}
\value{
A \code{list} of \code{numeric} vectors of ETG curve modeled signal corresponding to and equal
in length to retention time \code{t}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_icf}} workflow, not intended for standalone use.

Builds an Empirically-Transformed Gaussian curve for up to 5 peaks to be used in conjunction with error calculation function \code{\link{etg_penalty}}.
See \code{\link{chrom_icf}} for details.
}
\seealso{
\code{\link{chrom_icf}}, \code{\link{etg_penalty}}
}
\keyword{internal}
