% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{icf_GS}
\alias{icf_GS}
\title{Build a Gaussian model for one or more peaks}
\usage{
icf_GS(
  H,
  center,
  sigma,
  H2 = NA,
  center2 = NA,
  sigma2 = NA,
  H3 = NA,
  center3 = NA,
  sigma3 = NA,
  H4 = NA,
  center4 = NA,
  sigma4 = NA,
  H5 = NA,
  center5 = NA,
  sigma5 = NA,
  t
)
}
\arguments{
\item{H, H2, H3, H4, H5}{The height at peak maximum for 1-5 peaks.}

\item{center, center2, center3, center4, center5}{The retention time at peak maximum for 1-5 peaks.}

\item{sigma, sigma2, sigma3, sigma4, sigma5}{The standard deviation of the Gaussian curve for 1-5 peaks.}

\item{t}{A \code{numeric} vector of peak retention times.}
}
\value{
A \code{list} of \code{numeric} vectors of Gaussian curve modeled signal corresponding to and equal
in length to retention time \code{t}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_icf}} workflow, not intended for standalone use.

Builds a Gaussian curve for up to 5 peaks to be used in conjunction with error calculation function \code{\link{gs_penalty}}.
See \code{\link{chrom_icf}} for details.
}
\seealso{
\code{\link{chrom_icf}}, \code{\link{gs_penalty}}
}
\keyword{internal}
