% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{icf_plot}
\alias{icf_plot}
\title{Visualize the results of Non-Linear Least Squares Iterative Curve Fitting (ICF)}
\usage{
icf_plot(
  input,
  cols = "default",
  plabs = "default",
  plotsum = TRUE,
  plotind = FALSE,
  fillplot = TRUE,
  txtmax = TRUE,
  norm = FALSE,
  asprat = 0.71
)
}
\arguments{
\item{input}{The \code{data.frame} of data output from the \code{[["main_data"]]} element of \code{\link{chrom_icf}}.}

\item{cols}{Plot colour palette. Either \code{"default"} or a \strong{named} \code{character} vector of colours.
Names denote the following plot elements: original chromatogram (\code{"main"}), Gaussian (\code{"gs"}), EGH (\code{"egh"}),
EMG (\code{"emg"}), and ETG (\code{"etg"}) grouped models, unmodeled peaks (\code{"none"}),
and models for individual peaks rather than groups (\code{"indiv"}).}

\item{plabs}{Plot labels. Either \code{"default"} or a \strong{named} \code{character} vector of labels for the following elements:
main plot title (\code{"main"}), retention time (\code{"x"}), and signal (\code{"y"}).}

\item{plotsum}{A \code{logical} switch. Should summed models for peak groups be plotted? Defaults to \code{TRUE}.}

\item{plotind}{A \code{logical} switch. Should individual peak models be plotted? Defaults to \code{FALSE}.}

\item{fillplot}{A \code{logical} specifying whether peak models should be coloured/filled (\code{TRUE} by default).
When \code{FALSE}, only outlines of peak models are plotted.}

\item{txtmax}{A \code{logical} switch specifying whether to plot peak numbers at apices. Defaults to \code{FALSE}.
Only applies to plots of individual peak models (\code{plotind = TRUE}).}

\item{norm}{A \code{logical} specifying whether to normalize the chromatographic signal (y-axis) to a percentage scale (0 to 100).
Defaults to \code{FALSE}.}

\item{asprat}{Aspect ratio of the plot (defaults to \code{0.71}).}
}
\value{
A \code{ggplot}-class object containing the plot.
}
\description{
Visualizes the results of Non-Linear Least Squares Iterative Curve Fitting (ICF) of chromatographic data
obtained via \code{\link{chrom_icf}}.
}
\examples{
\dontrun{
#Get data and plot
icf_res <- lcqc:::wf_icf[["main_data"]]
icf_plot(icf_res)

#Normalized plot with models for individual peaks included
icf_plot(icf_res, plotind = TRUE, norm = TRUE)
}

}
\seealso{
\code{\link{chrom_icf}}
}
