% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{inf_refine}
\alias{inf_refine}
\title{Refine inflection points of peaks}
\usage{
inf_refine(sig, sdmin, apex, linfmrk, rinfmrk, orig_l, orig_r, bds_l, bds_r)
}
\arguments{
\item{sig}{Signal data (\code{numeric}).}

\item{sdmin}{Indices of second derivative minima (\code{numeric}).}

\item{apex}{Indices of peak maxima (\code{numeric}).}

\item{linfmrk, rinfmrk}{Indices of \strong{all} second derivative downcrosses (\code{numeric}, \code{linfmrk}) and upcrosses (\code{numeric}, \code{rinfmrk}) marking inflection points.}

\item{orig_l, orig_r}{Indices of \strong{current} left and right inflection points. Must each be equal in length to that of \code{apex}.}

\item{bds_l, bds_r}{Indices of peak start (\code{bds_l}) and end (\code{bds_r}) boundaries.}
}
\value{
A \code{list} with refined \code{$left} and \code{$right} inflection points.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Refines inflection points of peaks to be as close as possible to their Gaussian theoretical values.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}
}
\keyword{internal}
