% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{init_egh}
\alias{init_egh}
\title{Estimate initial Gaussian standard deviation and exponential time constant for EGH models}
\usage{
init_egh(A = NA, B = NA, frac = NA)
}
\arguments{
\item{A, B}{The \code{numeric} left and right (\code{A} and \code{B}, respectively) half-widths of the peak determined at \code{frac}.}

\item{frac}{The \code{numeric} peak height (signal) fraction (between 0 and 1) at which \code{A} and \code{B} were determined.}
}
\value{
A named \code{numeric} vector containing the estimated \code{"sigma"} and \code{"tau"} values.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_icf}} workflow, not intended for standalone use.

Estimates the standard deviation (sigma) of the Gaussian component and the time constant (tau) of the truncated exponential component
for the Exponential-Gaussian Hybrid (EGH) model from peak data. See \code{\link{chrom_icf}} for further details.
}
\seealso{
\code{\link{chrom_icf}}, \code{\link{icf_EGH}}, \code{\link{egh_penalty}}
}
\keyword{internal}
