% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{integ}
\alias{integ}
\title{Trapezoidal Rule integration of data}
\usage{
integ(input, vars = c("x", "y"), slnt = FALSE)
}
\arguments{
\item{input}{Either a \code{data.frame} or a \code{list} of \code{data.frame} objects containing \strong{xy} data to be integrated.}

\item{vars}{A \code{character} vector of length 2 specifying column names of \strong{x-} and \strong{y-}data included in \code{input}.
Defaults to \code{c("x","y")}.}

\item{slnt}{A \code{logical} (defaults to \code{FALSE}). Specifies whether information about the function output is printed in the console.}
}
\value{
A \code{numeric} vector of integrated peak areas, each element corresponding to that of \code{input}
(or a single \code{numeric} peak area if \code{input} is a \code{data.frame}).
}
\description{
Carries out integration of data using the Trapezoidal Rule.
See \code{\link{chrom_skim}} for further details.
}
\seealso{
\code{\link{chrom_skim}}
}
\keyword{internal}
