% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.even}
\alias{is.even}
\alias{is.odd}
\title{Find whether a number is even or odd}
\usage{
is.even(x)

is.odd(x)
}
\arguments{
\item{x}{Single numeric integer value.}
}
\value{
A TRUE/FALSE depending on whether the input is an odd (for \code{is.odd}) or even (for \code{is.even}) number.
}
\description{
Returns a logical values based on whether an input number is even or odd.
}
\examples{
is.even(1)
is.odd(3)
}
