% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcqc-package.R
\docType{package}
\name{lcqc-package}
\alias{lcqc}
\alias{lcqc-package}
\title{lcqc: Chromatographic Peak Detection, Analysis, and Reporting}
\description{
LCQC is used to automatically detect, process, and analyze chromatographic peaks for various metrics relevant to quality control of liquid chromatographic (HPLC) columns. Functions for baseline correction, chromatogram smoothing, peak detection and integration, non-linear iterative curve fitting (peak deconvolution), performance metric calculations, and automated reporting are included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Deniz-Koseoglu/lcqc}
  \item Report bugs at \url{https://github.com/Deniz-Koseoglu/lcqc/issues}
}

}
\author{
\strong{Maintainer}: Deniz Can Koseoglu \email{denizcan.koseoglu@gmail.com} (\href{https://orcid.org/0000-0002-1569-9788}{ORCID})

}
\keyword{internal}
