% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_defs}
\alias{render_defs}
\alias{lcqc_render}
\title{Render LCQC data into a concise HPLC column performance report}
\usage{
render_defs()

lcqc_render(
  tpl,
  cplot,
  tpars,
  spec,
  clogo = NA,
  sym = NA,
  ext = NA,
  mets = c("EP", "AH", "Tf", "As"),
  exmets = c("Linear Velocity", "Packing Porosity", "Flow Resistance", "Permeability"),
  add_tpa = FALSE,
  expath = getwd(),
  asprat = 0.4,
  pnms = NA,
  pconcs = NA,
  add_pnums = TRUE,
  fontsize = 10,
  which_pks = "all"
)
}
\arguments{
\item{tpl}{The output of function \code{\link{chrom_tplate}} containing theoretical plates and other metrics.}

\item{cplot}{A \code{ggplot} of the chromatogram to be included in the report.
Usually obtained via the \code{\link{chrom_skim}} or \code{\link{chrom_icf}} workflows.}

\item{tpars}{A \strong{named} \code{character} vector specifying various general text to be included in the report.
Values may include (also run \code{\link{render_defs}} for examples):
\enumerate{
\item Column serial number \code{"sn"}
\item Column part number \code{"pn"}
\item Short column description \code{"desc"}
\item Batch number \code{"bn"}
\item Document number \code{"dnum"}
\item Operator name and surname \code{"oper"}
\item Mobile phase specification \code{"mp"}
\item Shipment phase \code{"sp"} (which solvent is the column shipped in?)
\item Back pressure \code{"bp"}
\item Flow rate \code{"flow"}
\item Column temperature \code{"temp"}
\item Injection volume \code{"inj"}
\item Analyte concentration unit \code{"unit_c"}
\item Timestamp for the report genesis \code{"tform"} (\code{"auto"} by default)
}}

\item{spec}{A \strong{named} \code{list} of QC specifications to display in the report. Names \strong{must match}
the elements of \code{mets}.}

\item{clogo}{A \code{character} filepath to a custom \strong{.JPG} logo to include in the report header.
A placeholder is displayed if no logo is provided.}

\item{sym}{\strong{Optional} asymmetry metrics and Total Peak Analysis output from function \code{\link{chrom_asym}}.}

\item{ext}{\strong{Optional} additional column-specific metrics output from function \code{\link{chrom_addmets}}.}

\item{mets}{A \code{character} vector of theoretical plate (\emph{N}) and asymmetry metrics to display in the report.
Possible metrics are: European Pharmacopoeia \emph{N} (\code{"EP"}), Area-Height \emph{N} (\code{"AH"}),
sigma-5 \emph{N} (\code{"S5"}), Exponentially-Modified Gaussian \emph{N} (\code{"EMG"}), Asymmetry Factor (\code{"As"}),
and/or USP/EP Tailing Factor (\code{"Tf"}).}

\item{exmets}{\strong{Optional} \code{character} vector of \strong{additional} column-specific metrics to display
at the bottom of the first page. \strong{Only relevant if} \code{ext} is provided.
Possible values are \code{"Linear Velocity"}, \code{"Packing Porosity"}, \code{"Flow Resistance"}, and/or \code{"Permeability"}.}

\item{add_tpa}{A \code{logical} switch indicating whether to add TPA plots to the second page of the report
(defaults to \code{FALSE}).}

\item{expath}{A \code{character} directory path into which to export the .PDF report.}

\item{asprat}{Aspect ratio of the chromatogram plot (\code{0.4} by default).}

\item{pnms}{Either \code{NA} (default) or a \code{character} vector of peak names to include in the report.
Must be equal in length to the total number of peaks in the input data, or to \code{which_pks} if these are specified.}

\item{pconcs}{Either \code{NA} (default) or a vector of concentrations corresponding to each peak included in the data,
or to those specified in \code{which_pks}.}

\item{add_pnums}{A \code{logical} switch specifying whether to add peak index numbers to the PDF report \strong{when \code{pnms} are given}.
Defaults to \code{TRUE}.}

\item{fontsize}{The font size to use for main text of the report (\code{10} by default).}

\item{which_pks}{Either \code{"all"} (default) or a \code{numeric} vector of peak indices to include in the report.}
}
\value{
The column performance report .PDF file is exported to the directory specified in \code{expath}.
}
\description{
Compiles key data obtained from the \pkg{lcqc} workflow into a
modular HPLC Column Performance Report in .PDF format. The function utilizes
the \strong{Quarto} framework for R.
}
\examples{
\dontrun{
#View available definitions
render_defs()

#Render .PDF report into the working directory
lcqc_render(tpl = lcqc:::wf_tplate,
cplot = lcqc:::wf_ints[["plot"]], #ALWAYS remove title and legend from plot
sym = lcqc:::wf_asyms,
ext = lcqc:::wf_perfmets,
which_pks = 4:7,
pnms = c("Naphthalene", "Anthracene", "Caffeine", "Benzene"),
pconcs = c(250, 500, 250, 500),
mets = c("EP", "AH", "Tf", "As"),
exmets = c("Linear Velocity", "Packing Porosity", "Flow Resistance", "Permeability"),
add_tpa = TRUE,
spec = list(EP = rep(">4000", 4),
           AH = rep(">4000", 4),
           Tf = rep("0.8-1.2",4),
           As = rep("0.8-1.2",4)),
tpars = c(sn = "C18-2024000001",
         pn = "L-C182546510",
         desc = "AC-C18 Column\n(250x4.6 mm, 5 micron, 10 nm)",
         bn = "20240000001",
         dnum = "LC-C18-20240000001",
         oper = "Deniz Can Koseoglu",
         mp = "70:30 MeOH:Water",
         sp = "80:20 (volumetric) IPA:Water",
         bp = "126 bar",
         flow = "1.0 mL/min",
         temp = "30 degC",
         inj = "20 microL",
         unit_c = "ppm",
         tform = "default"),
expath = getwd(),
asprat = 0.4,
add_pnums = TRUE,
fontsize = 10)
}

}
\references{
Allaire, J.J., Teague, C., Scheidegger, C., Xie, Y., Dervieux, C., (2024), \emph{Quarto, version 1.4}, DOI: \url{https://www.doi.org/10.5281/zenodo.5960048}.
}
\seealso{
\code{\link{render_defs}}, \code{\link{chrom_tplate}}, \code{\link{chrom_asym}}, \code{\link{chrom_addmets}},
\code{\link{chrom_skim}}
}
