% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{linspline}
\alias{linspline}
\alias{cubspline}
\title{Find roots via linear or non-linear interpolation}
\usage{
linspline(x, y, y0 = 0, verbose = TRUE)

cubspline(f, y0 = 0, verbose = TRUE)
}
\arguments{
\item{x}{The known x values (only for \code{linspline}).}

\item{y}{The known y values (only for \code{linspline}).}

\item{y0}{The y value for which an x value is to be found (defaults to 0, indicating the root is to be found).}

\item{verbose}{A logical which determines whether the results are plotted using \code{plot()}.}

\item{f}{Cubic spline results obtained from \code{stats::splinefun}.}
}
\value{
A numeric vector of intersection point(s) along the x-axis.
}
\description{
By default, function \code{linspline} finds the zero-crossing (i.e. root) of a linear interpolation function.
Similarly, \code{cubspline} finds the root of a cubic spline (output from \code{stats::splinefun}).
}
\examples{
set.seed(0)
x <- 1:10 + runif(10, -0.1, 0.1)
y <- rnorm(10, 3, 1)
f <- stats::splinefun(x, y, method = "fmm")
linspline(x, y, y0 = 2.85, verbose = FALSE)
cubspline(f, y0 = 2.85, verbose = FALSE)
}
\author{
Zheyuan Li (see also \href{https://stackoverflow.com/questions/52655729/get-x-value-given-y-value-general-root-finding-for-linear-non-linear-interpol}{here})
}
