% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{markmerge}
\alias{markmerge}
\title{Merge peak markers from original and smoothed derivatives}
\usage{
markmerge(marks, smooth_marks = NA, max_w = NA, minsim = 3)
}
\arguments{
\item{marks}{The \code{\link{peakmark}} output (i.e. compliant and non-compliant peak markers) derived from original (unsmoothed) derivatives.}

\item{smooth_marks}{Either \code{NA} or a \code{data.frame} output from \code{\link{peakmark}} using smoothed derivatives. When \code{NA},
\code{marks} are returned as a \code{list} (see \strong{Value}).}

\item{max_w}{Additional maximum width parameter (in point indices) to use for second derivative maxima/minima when \code{minsim} fails.
Defaults to \code{NA}. See \strong{Details}.}

\item{minsim}{The distance (in point indices) within which to look for matching markers between \code{marks} and \code{smooth_marks}.
Defaults to ±3 points.}
}
\value{
A named \code{list} where each element contains a specific type of peak marker (name suffixes \code{_minima}, \code{_maxima}, \code{_downcr},
\code{_upcr}) for signal, first derivatives, or second derivatives (prefixes \code{sig}, \code{FD}, or \code{SD}).
Non-compliant markers are additionally suffixed with \code{_nc}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Merges the results of \code{\link{peakmark}} into a final list of peak markers based on their similarity.
}
\details{
The function tests whether a peak marker of a given type in \code{marks} has an equivalent in \code{smooth_marks} that is within ±\code{minsim}.
Where this is true, the rounded \code{mean} value of both markers is taken as the final value.
Where \code{minsim} criterion fails and the marker type is a second derivative extreme (minimum or maximum), 1/4 of \code{max_w}
(representing the maximum smooth width - a product of smoothing points and passes) is applied as an additional criterion.
If \code{max_w} is \code{NA} or <10, it is set to 10 points. Else, the presence of \strong{signal} maximum or minimum within ±\code{max_w} is checked
when the marker in question is a second derivative minimum or maximum, respectively.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{chrom_smooth}}, \code{\link{peakmark}}, \code{\link{peakfind}}
}
\keyword{internal}
