% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{noise_calc}
\alias{noise_calc}
\title{Get noise value from chromatographic signal}
\usage{
noise_calc(sig, sder, sig_thres, sd_thres)
}
\arguments{
\item{sig}{Signal \code{numeric} vector.}

\item{sder}{Second derivatives corresponding to \code{sig}, e.g. obtained via \code{\link{chrom_deriv}}.}

\item{sig_thres}{A single \code{numeric} value of the signal amplitude limit (from \code{\link{chrom_amplim}}).}

\item{sd_thres}{A \code{numeric} vector of length 2 containing \strong{second derivative} peak detection thresholds (from \code{\link{chrom_derlims}}).}
}
\value{
A single \code{numeric} noise value for \code{sig}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Calculates the noise component of the S/N ratio, e.g. for use in \code{\link{rej_filter}}.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{rej_filter}}, \code{\link{chrom_derlims}},
\code{\link{chrom_deriv}}
}
\keyword{internal}
