% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_plot.R
\name{noise_plot}
\alias{noise_plot}
\title{Visualise noise data}
\usage{
noise_plot(
  noise,
  thres,
  plot_title = "Peak Recognition Thresholds",
  y_lab = "Noise",
  asprat = 0.71
)
}
\arguments{
\item{noise}{A \code{numeric} vector containing noise data.}

\item{thres}{A \code{numeric} vector of length 2 containing the lower and upper noise thresholds.}

\item{plot_title}{The \code{character} plot title.}

\item{y_lab}{The \code{character} y-axis title.}

\item{asprat}{Aspect ratio of the plot.}
}
\value{
A \code{ggplot}-class plot object.
}
\description{
Create a summary plot of noise data, usually derived from \code{\link{chrom_derlims}}.
}
\examples{
#Get noise and threshold data
\dontrun{
dt <- lcqc:::wf_detpeaks[["results"]]
nsdt <- dt[["Derivative_Noise"]][["SD_noise"]]
thrdt <- dt[["Derivative_Limits"]][["SD"]]
noise_plot(nsdt, thrdt)
}

}
\seealso{
\code{\link{chrom_derlims}}
}
