% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{pdrop}
\alias{pdrop}
\title{Derive Perpendicular Drop line}
\usage{
pdrop(str, end)
}
\arguments{
\item{str, end}{A 2-row \code{data.frame} containing key information about parent and child peak \strong{start} and \code{end} boundaries as output from
\code{\link{dtprep}}\code{$type_df$starts} and \code{\link{dtprep}}\code{$type_df$ends}, respectively.}
}
\value{
A \code{data.frame} containing original data indices (\code{"ind"}), retention time (\code{"x"}), baseline-corrected signal (\code{"y"}),
peak group ID (\code{"group"}), those of the parent and child peak (\code{"peak1"} and \code{"peak2"}), the skim type (one of \code{"PD"}),
and the combined ID used for plotting via \code{\link{integ_plot}} (\code{"combo_id"}).
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Compiles data necessary for plotting a Perpendicular Drop line via \code{\link{integ_plot}}.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{integ_plot}}
}
\keyword{internal}
