% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icf_helpers.R
\name{peak_hw}
\alias{peak_hw}
\title{Calculate peak half-widths at any peak height fraction(s)}
\usage{
peak_hw(
  input,
  vars = c("x", "y"),
  accmax = NULL,
  frac = 0.15,
  resolved = TRUE,
  slnt = FALSE
)
}
\arguments{
\item{input}{A \code{data.frame} containing retention time and signal values (specified in \code{vars}) for the peak to be processed.}

\item{vars}{A \code{character} vector of retention time (x-axis) and signal (y-axis) column names. Defaults to \code{c("x","y")}.}

\item{accmax}{A \code{numeric} vector (or \code{data.frame}) containing retention time (x-axis) and signal (y-axis) of the peak maximum,
\strong{in that order}.}

\item{frac}{A \code{numeric} vector of peak height fractions (between 0.00 and 1.00) to calculate peak width and half-widths at.}

\item{resolved}{A \code{logical} indicating whether the peak to be assessed is baseline-resolved (\code{TRUE} by default).}

\item{slnt}{A \code{logical} switch determining whether information about results is printed in the console.}
}
\value{
A 4-column \code{data.frame} containing the peak height fraction(s) specified in \code{frac} (\code{"frac"}) with
associated total width (\code{"W"}), left half-width (\code{"A"}), and right half-width (\code{"B"}).
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_icf}} workflow, not intended for standalone use.
Calculates peak width and half-widths are specified peak height fractions (from 0 to 1).
}
\seealso{
\code{\link{chrom_icf}}, \code{\link{chrom_tplate}}, \code{\link{chrom_asym}}, \code{\link{chrom_tpa}}, \code{\link{chrom_res}}
}
\keyword{internal}
