% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind.R
\name{peakfind}
\alias{peakfind}
\title{A workflow to identify baseline-resolved, fused, shoulder, and round chromatographic peaks}
\usage{
peakfind(
  marks,
  smooth_marks = NA,
  rtime,
  sig,
  fder,
  sder,
  max_w = NA,
  liftoff = 0,
  touchdown = 0.5,
  sig_thres,
  fd_thres,
  sd_thres,
  min_ht = NA,
  w_rej = NA,
  pa_rej = NA,
  sn_rej = NA,
  rej_logic = rep("OR", 2),
  bnch = TRUE,
  refine_infs = TRUE,
  fchrom = NA
)
}
\arguments{
\item{marks}{Initial peak markers from \strong{original unsmoothed} data output from function \code{\link{markmerge}}.}

\item{smooth_marks}{(\strong{Optional}) initial peak markers from \strong{smoothed} data output from function \code{\link{markmerge}}.}

\item{rtime}{A \code{numeric} vector of retention time data.}

\item{sig}{A \code{numeric} vector of chromatographic signal data.}

\item{fder, sder}{First and second derivative data as \code{numeric} vectors. Must be of equal length to \code{rtime} and \code{sig}.}

\item{max_w}{The maximum smooth width (in point indices). Set to 10 where a lower value is given.}

\item{liftoff, touchdown}{Single \code{numeric} parameters (between 0 and 100) required by the ApexTrack algorithm (see \code{\link{apexbnds}}).}

\item{sig_thres}{Signal amplitude limit (threshold) for peak detection (see \code{chrom_amplim}).}

\item{fd_thres, sd_thres}{Both are \code{numeric} vectors of length 2 containing the first and second derivative peak detection thresholds
(see \code{\link{chrom_derlims}}).}

\item{min_ht}{Height peak \strong{pre-}-filtering criterion (\code{numeric}). Peaks of lower height are removed. Defaults to \code{NA} (no filter applied).}

\item{w_rej}{Inflection point width peak \strong{post}-filtering criterion (\code{numeric}). Defaults to \code{NA}.}

\item{pa_rej}{Peak \strong{post}-filtering criterion for inflection point peak area (\code{numeric}). Defaults to \code{NA}.}

\item{sn_rej}{Signal-to-noise ratio (S/N) peak \strong{pre}-filtering criterion. Defaults to \code{NA}.}

\item{rej_logic}{Logic to use for peak filtering via \code{w_rej}, \code{pa_rej}, \code{min_ht}, and/or \code{sn_rej}. One of \code{"OR"} or \code{"AND"}. See \code{\link{rej_filter}} for more information.}

\item{bnch}{A \code{logical} indicating whether peak bunching should be detected and mitigated (see \code{\link{detect_bunch}}).}

\item{refine_infs}{A \code{logical} indicating whether detected inflection points should be refined to be as close as possible to their Gaussian theoretical values
(see \code{\link{inf_refine}}).}

\item{fchrom}{Either \code{NA} (default) or a \code{numeric} value of critical width to use for FastChrom baseline correction.
Applies to resolved peaks only. See \code{\link{fastchrom_bline}} for details.}
}
\value{
A list of length 2 containing the following elements:
\describe{
\item{results}{A \code{data.frame} of peak groups, peak types (\code{ptype}), peak boundary types (\code{lb_type}, and \code{rb_type}),
and extent markers (as data indices) with their corresponding retention times. Markers include second derivative-based maxima (\code{_dermax}),
signal maxima (\code{_sigmax}), final maxima (\code{_finmax}), start and end boundaries (\code{_starts} and \code{_ends}),
inflection points (\code{_linf} and \code{_rinf}), and upslope points (\code{_lups} and \code{_rups}).
Various peak metrics used for filtering (e.g. via \code{\link{rej_filter}}) are also included, such as the peak area (\code{"top_pa"}),
inflection point width (\code{"inf_wd"}), and S/N ratio (\code{"sn_ratio"}).}
\item{information}{A \code{character} string summarising information about the number and type(s) of detected peaks.}
}
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

This function forms a complete workflow for peak identification and uses various component functions (consult the \strong{See Also} section).
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{markmerge}}, \code{\link{sdmin_infchk}}, \code{\link{detect_bunch}}, \code{\link{noise_calc}},
\code{\link{rej_filter}}, \code{\link{det_infs}}, \code{\link{base_raw}}, \code{\link{find_fused}}, \code{\link{base_fused}}, \code{\link{find_shld}},
\code{\link{ptype_class}}, \code{\link{corresp}}, \code{\link{grp_pks}}, \code{\link{inf_refine}}, \code{\link{find_ups}}, \code{\link{fastchrom_bline}}
}
\keyword{internal}
