% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{peakmark}
\alias{peakmark}
\title{Locate initial peak markers}
\usage{
peakmark(
  sig,
  fder,
  sder,
  FD_thres,
  SD_thres,
  sig_thres,
  mpts = 3,
  crosspts = 2,
  sens = c(0.1, 1, 1)
)
}
\arguments{
\item{sig}{A \code{numeric} vector of signal data.}

\item{fder}{First derivatives corresponding to \code{sig}.}

\item{sder}{Second derivatives corresponding tp \code{sig}.}

\item{FD_thres}{A \code{numeric} vector of length 2 containing minimum and maximum \strong{first} derivative noise ranges (see \code{\link{chrom_derlims}}).}

\item{SD_thres}{A \code{numeric} vector of length 2 containing minimum and maximum \strong{second} derivative noise ranges (see \code{\link{chrom_derlims}}).}

\item{sig_thres}{A single \code{numeric} value of the signal amplitude limit used for peak detection (see \code{\link{chrom_amplim}}).}

\item{mpts}{Number of points surveyed on either side to confirm potential maxima and minima.}

\item{crosspts}{Number of points surveyed on either side to confirm zero-crossings (upcrosses and downcrosses).}

\item{sens}{A \code{numeric} vector of length 3 containing various sensitivity factors for fine-tuning (\code{c(0.1,1,1)} by default. Also see \strong{Details}).}
}
\value{
A \code{data.frame} containing indices of detected compliant and non-compliant (suffix \code{"_nc"}) peak markers.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Estimates peak markers from a combination of signal data as well as first and second derivatives.
Estimated markers include minima and maxima for signal and all derivatives. For first and second derivatives only, zero upcrosses and downcrosses are also found.
Both \strong{compliant} and \strong{non-compliant} markers are estimated.
}
\details{
Initially, maxima/minima and up/downcrosses are calculated taking \code{mpts} and \code{crosspts} into account, respectively.
From these total markers, compliant ones are extracted using several criteria. First, \strong{second derivative minima} (representing peak apices)
are filtered such that they are above the signal amplitude limit and below the minimum second derivative threshold multiplied by a sensitivity factor
(i.e. \code{> sig_thres & < sens[2]*min(SD_thres)}). Thus, a higher value of \code{sens[2]} leads to a more selective filter and decreases the number of compliant minima.
For detecting compliant \strong{signal} minima, only the amplitude limit is considered and \code{sens[1]} is used as the sensitivity factor.

For maxima, a similar procedure is followed. Signal maxima are considered compliant if \code{> sig_thres & < min(SD_thres)}.
For first and second derivatives, this condition is revised to \code{> sens[2]*max(FD_thres) | < sens[2]*min(FD_thres)} and
\code{> sens[3]*max(SD_thres) | < sens[3]*min(SD_thres)}, respectively.

For zero downcrosses and upcrosses alike, final indices are chosen such that all values within \code{±crosspts} do not constitute additional crosses,
while the final chosen value is closest to zero among those surveyed.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{chrom_deriv}}, \code{\link{chrom_derlims}}, \code{\link{chrom_amplim}}
}
\keyword{internal}
