% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{ptype_class}
\alias{ptype_class}
\title{Classify peak apices based on various peak markers and boundary types}
\usage{
ptype_class(bnds, infs, sdmin, sigmax, fdcross, fdcross_nc)
}
\arguments{
\item{bnds}{A \strong{named} \code{numeric} vector of \strong{all} peak boundary indices. Possible types are: \code{"B"} (baseline-resolved),
\code{"F"} (fused), \code{"S"} (shoulder), or \code{"R"} (round).}

\item{infs}{A \code{numeric} vector of inflection point indices.}

\item{sdmin}{Indices of second derivative minima (\code{numeric}).}

\item{sigmax}{Indices of signal maxima (\code{numeric}).}

\item{fdcross}{Indices of \strong{compliant} first derivative downcrosses (\code{numeric}).}

\item{fdcross_nc}{\strong{Non-compliant} equivalent of \code{fdcross}.}
}
\value{
A \code{character} vector of peak apex types including any of: \code{"B"}, \code{"F"}, \code{"S"}, and/or \code{"R"}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Classifies peak apices into one of four categories (see \strong{Arguments} and \strong{Value}).
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}
}
\keyword{internal}
