% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{raw_infs}
\alias{raw_infs}
\title{Detect closest inflection point either side of a peak}
\usage{
raw_infs(peak_val, outlim = NA, sdcrosses, sdcrosses_nc, which_side)
}
\arguments{
\item{peak_val}{The \code{numeric} index of a peak apex.}

\item{outlim}{The limiting index for detection. Inflection points are detected within \code{sort(peak_val:outlim)}.}

\item{sdcrosses, sdcrosses_nc}{Compliant and (optionally) non-compliant \code{numeric} indices of second derivative upcrosses and downcrosses.}

\item{which_side}{Which side of the peak to use for inflection point search? One of: \code{"left"} or \code{"right"}.}
}
\value{
A single \code{numeric} inflection point or \code{NA} if such a point is not found.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Detects a single closest inflection point either side of a peak apex.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}, \code{\link{det_infs}}
}
\keyword{internal}
