% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers_parsers.R
\name{read_shim}
\alias{read_shim}
\title{Read Shimadzu LabSolutions GC-FID, GC-MS, and HPLC data from ASCII file}
\usage{
read_shim(
  file,
  ptable = TRUE,
  simtable = TRUE,
  pnames = TRUE,
  pcas = TRUE,
  metadata = TRUE,
  mode = "fid",
  sep = "auto",
  decsep = "auto",
  fix_names = TRUE,
  fil_cols = TRUE,
  cols = "default",
  chromcols = "default",
  rm_dups = FALSE,
  unify_cols = NA,
  trange = NA
)
}
\arguments{
\item{file}{The \code{character} file path to a compatible Shimadzu ASCII.}

\item{ptable}{A \code{logical} specifying whether the peak table is extracted. Defaults to \code{TRUE}.}

\item{simtable}{A \code{logical} specifying whether the GC-MS similarity table is extracted. Only valid when \code{mode = "gcms"}.
Defaults to \code{TRUE}.}

\item{pnames}{A \code{logical} specifying whether peak names are retrieved (where available). \code{TRUE} by default.}

\item{pcas}{A \code{logical} specifying whether CAS numbers corresponding to peak names are retrieved.
Only valid when \code{mode = "gcms"}. \code{TRUE} by default.}

\item{metadata}{A \code{logical} specifying whether metadata is retrieved (defaults to \code{TRUE}).}

\item{mode}{A \code{character} value of the type of chromatogram to be read. One of: \code{"fid"} (GC-FID; default),
\code{"gcms"} (GC-MS), or \code{"lc"} (HPLC).}

\item{sep, decsep}{A single \code{character} indicating the \strong{column} and \strong{decimal} (\code{"decsep"}) separator
of input ASCII files. The default value \code{"auto"} attempts to determine this automatically.}

\item{fix_names}{A \code{logical} specifying whether column names should be made syntactically correct (\code{TRUE} by default).}

\item{fil_cols}{A \code{logical} specifying whether to truncate the peak table to only include the most relevant columns
(\code{TRUE} by default).}

\item{cols, chromcols}{Either \code{"default"} of a \code{character} vector of column names to extract from the \strong{peak table}
(\code{cols}) or \strong{chromatogram} (\code{chromcols}).}

\item{rm_dups}{A \code{logical} specifying whether duplicates should be removed from the GC-MS peak table.
Only valid when \code{mode = "gcms"}. Defaults to \code{FALSE} by default.}

\item{unify_cols}{Either \code{NA} (default) or an \strong{ordered} \code{character} vector of \strong{length 8} containing
universal column names to assign for all function output and corresponding to the following variables:
peak number, retention time, start time, end time, area, height, area/height ratio, and compound name.}

\item{trange}{Either \code{NA} (default) or a \code{numeric} vector of length \strong{1} or \strong{2} specifying
the \strong{lower} and/or \strong{upper} retention time limit(s) used to truncate the chromatogram and peak table.}
}
\value{
A named \code{list} containing various data extracted from the ASCII file. These include \code{metadata},
\code{chromatogram} (one or more), \code{ptable} (peak table), \code{simtable} (MS similarity table), \code{pnames}
(a \code{character} vector of peak names), and \code{pcas} (a \code{character} vector of corresponding CAS numbers).
When \code{mode = "gcms"} and Selected Ion Monitoring (SIM) chromatograms are present, they are all extracted alongside
the Total Ion Current (TIC) chromatogram as separate list elements. Similarly, all wavelength-specific LC chromatograms
are extracted separately when \code{mode = "lc"}.
}
\description{
Reads ASCII files originating from Shimadzu LabSolutions and
GCMS Solutions software and extracts the chromatogram, peak table, metadata,
and MS similarity search table. Modified and extended from function
\code{\link[chromConverter]{read_shimadzu}} from the \pkg{chromConverter} package.
}
\examples{
\dontrun{
#Get data paths of external Shimadzu ASCII data included with lcqc
dtp <- ext_lcqc()

#Read FID chromatogram
fidc <- read_shim(dtp[grep("exgc_",dtp)[1]], mode = "fid")

#Read GCMS chromatogram and peak table
msc <- read_shim(dtp[grep("exgcms_",dtp)[1]], mode = "gcms")

#Read HPLC chromatogram
lcc <- read_shim(dtp[grep("exlc_",dtp)[1]], mode = "lc")
}

}
\references{
Bass, E. (2022), \emph{chromConverter}, available at: \url{https://cran.r-project.org/package=chromConverter} (accessed 25.04.2024).
}
\seealso{
\code{\link{ext_lcqc}}
}
