% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfind_helpers.R
\name{rej_filter}
\alias{rej_filter}
\title{Filter peaks based on various chromatography-relevant criteria}
\usage{
rej_filter(
  rtime = NA,
  sig,
  apices,
  linfs = NA,
  rinfs = NA,
  crit = c(ht = NA, wd = NA, pa = NA, sn = NA),
  noise = NA,
  logic = "OR"
)
}
\arguments{
\item{rtime}{Retention time \code{numeric} vector.}

\item{sig}{Signal \code{numeric} vector.}

\item{apices}{Indices of peak apices along \code{rtime} and \code{sig}.}

\item{linfs, rinfs}{Indices of left and right inflection points.}

\item{crit}{A \strong{named} \code{numeric} vector of criteria to filter by, including: height (\code{ht}), inflection point width (\code{wd}),
inflection point area (\code{pa}), and/or S/N ratio (\code{sn}).}

\item{noise}{The noise component of the S/N ratio, calculated by the \code{\link{noise_calc}} function.}

\item{logic}{The logic used to combine different criteria. One of: \code{"OR"} or \code{"AND"}.}
}
\value{
A list of length 2:
\describe{
\item{results}{Vector of \code{logical} values equal in length to \code{apices}. \code{TRUE} marks peaks for removal.}
\item{info}{Information about the distribution of filtered peaks based on all relevant criteria.}
}
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_detect}} workflow, not intended for standalone use.

Filters detected peaks based on any combination of signal-to-noise ratio (S/N), height, as well as inflection point peak area and/or width.
}
\seealso{
\code{\link{chrom_detect}}, \code{\link{peakfind}}
}
\keyword{internal}
