% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the LCQC Shiny Application}
\usage{
run_app()
}
\value{
A shiny app object. The function starts the Shiny application and
does not return a value directly to the R session until the app is closed.
}
\description{
This function launches the LCQC Shiny application packaged within the 'lcqc' package.
}
\examples{
\dontrun{
# To run the application after installing the package:
lcqc::run_app()
}
}
