% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simlc1}
\alias{simlc1}
\title{Simulated LC Chromatogram (example 1)}
\format{
\subsection{\code{simlc1}}{

A data frame with 3000 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
simlc1
}
\description{
A simulated example two column HPLC chromatogram (1 of 7)
}
\keyword{datasets}
