% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simlc3}
\alias{simlc3}
\title{Simulated LC Chromatogram (example 3)}
\format{
\subsection{\code{simlc3}}{

A data frame with 3000 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
simlc3
}
\description{
A simulated example two column HPLC chromatogram (3 of 7)
}
\keyword{datasets}
