% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simlc4}
\alias{simlc4}
\title{Simulated LC Chromatogram (example 4)}
\format{
\subsection{\code{simlc4}}{

A data frame with 3000 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
simlc4
}
\description{
A simulated example two column HPLC chromatogram (4 of 7)
}
\keyword{datasets}
