% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simlc5}
\alias{simlc5}
\title{Simulated LC Chromatogram (example 5)}
\format{
\subsection{\code{simlc5}}{

A data frame with 3000 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
simlc5
}
\description{
A simulated example two column HPLC chromatogram (5 of 7)
}
\keyword{datasets}
