% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simlc6}
\alias{simlc6}
\title{Simulated LC Chromatogram (example 6)}
\format{
\subsection{\code{simlc6}}{

A data frame with 3000 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
simlc6
}
\description{
A simulated example two column HPLC chromatogram (6 of 7)
}
\keyword{datasets}
