% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simlc7}
\alias{simlc7}
\title{Simulated LC Chromatogram (example 7)}
\format{
\subsection{\code{simlc7}}{

A data frame with 3000 rows and 2 columns
\describe{
\item{Time}{Retention time of chromatogram}
\item{Signal}{Signal of chromatogram}
}
}
}
\usage{
simlc7
}
\description{
A simulated example two column HPLC chromatogram (7 of 7)
}
\keyword{datasets}
