% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{skimcomp}
\alias{skimcomp}
\title{Compile Perpendicular Drop or Tangent/Exponential/Gaussian skim data for trapezoidal integration}
\usage{
skimcomp(input, integ_data)
}
\arguments{
\item{input}{A \code{data.frame} containing the row-bound results of skimming functions (any of \code{\link{tskim}}, \code{\link{exskim}},
\code{\link{gskim}}, and/or \code{\link{pdrop}}) for all peaks to be integrated.}

\item{integ_data}{A \code{list} of \code{data.frame} objects, each containing data indices (\code{"ind"}), retention time (\code{"x"}),
and baseline-corrected signal (\code{"y"}) for an individual (often skimmed) peak to be used for integration.}
}
\value{
A \code{list} of \code{data.frame} objects, each containing the necessary xy-coordinates (retention time and baseline-corrected signal)
for an individual peak to be used for trapezoidal integration via \code{\link{integ}}.
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Compiles data for integration via the Trapezoidal Rule as part of the workflow. See \code{\link{chrom_skim}} for further details.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{tskim}}, \code{\link{exskim}}, \code{\link{gskim}}, \code{\link{pdrop}}, \code{\link{integ}}
}
\keyword{internal}
