% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slope}
\alias{slope}
\alias{intercept}
\title{Slope and intercept of a linear function}
\usage{
slope(x, y)

intercept(x, y)
}
\arguments{
\item{x}{Numeric vector of known x values.}

\item{y}{Numeric vector of Known y values.}
}
\value{
A numeric value of the linear gradient/slope.
}
\description{
Functions to calculate the \code{slope} and \code{intercept} of a line.
}
\examples{
with(CO2, slope(conc, uptake))
with(CO2, intercept(conc, uptake))
}
\seealso{
\url{https://stackoverflow.com/questions/66771929/intercept-and-slope-functions-in-r}
}
