% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{smooth_whit}
\alias{smooth_whit}
\title{Weighted (Whittaker-Henderson) smoother}
\usage{
smooth_whit(input, lambda, wts = rep(1, length(input)))
}
\arguments{
\item{input}{Input numeric vector containing signal for smoothing.}

\item{lambda}{Smoothing parameter (defaults to \code{1600}). Degree of smoothing increases with value.}

\item{wts}{Vector of relative weights, must equal \code{input} in length. Defaults to \code{1}.}
}
\value{
Numeric vector containing the smoother signal.
}
\description{
Smooth a signal with a finite difference penalty of order 2. For more details, see \strong{References}. Modified from package \pkg{ptw}.
}
\examples{
smooth_whit(lcqc::simlc1[,"Signal"], lambda = 1600)

}
\references{
Eilers, P.H.C. (2003), 'A perfect smoother', \emph{Analytical Chemistry} \strong{75}, pp. 3631-3636.

Eilers, P.H.C. (2004), 'Parametric Time Warping', \emph{Analytical Chemistry} \strong{76} (2), pp. 404-411.
}
\seealso{
\code{\link{bline_als}}
}
