% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_helpers.R
\name{tskim}
\alias{tskim}
\title{Front or tail Tangent Skim on a pair of peaks}
\usage{
tskim(child, cmax, skim_type)
}
\arguments{
\item{child}{A \code{data.frame} containing the \code{\link{dtprep}} \code{$grp_blines} output with \strong{child} peak data.}

\item{cmax}{A \strong{single row} \code{data.frame} containing \strong{child} peak data output from \code{\link{dtprep}}\code{$type_df$maxes}.}

\item{skim_type}{Type of skim to apply. One of \code{"front"} or \code{"tail"}.}
}
\value{
A \code{data.frame} containing original data indices (\code{"ind"}), retention time (\code{"x"}), baseline-corrected signal (\code{"y"}),
peak group ID (\code{"group"}), those of the parent and child peak (\code{"peak1"} and \code{"peak2"}), the skim type (one of \code{"TF"} or
\code{"TT"} for front and tail skims, respectively), and the combined ID used for plotting via \code{\link{integ_plot}} (\code{"combo_id"}).
}
\description{
\strong{This function is not exported}. Part of the \code{\link{chrom_skim}} workflow, not intended for standalone use.

Carries out a Tangent Skim on a suitable parent-child pair of chromatographic peaks.
See \code{\link{chrom_skim}} for further details.
}
\seealso{
\code{\link{chrom_skim}}, \code{\link{integ_plot}}
}
\keyword{internal}
