% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_thres.R
\name{z_thres}
\alias{z_thres}
\title{Thresholding based on z-scores}
\usage{
z_thres(y, lag, threshold, influence)
}
\arguments{
\item{y}{Numeric vector of equi-spaced y-values. Length needs to be \code{>=lag+2}.}

\item{lag}{Number of observations to base average and standard deviation on (e.g. 5).}

\item{threshold}{The z-score (factor to multiply standard deviation of the moving average by) at which the algorithm signals the presence of positive or negative peaks (e.g. 3.5).}

\item{influence}{Factor between 0 and 1 denoting the relative influence/weight that new data points have on the calculation of moving average and moving standard deviation (e.g. 0.5).}
}
\value{
A list of 4 elements:
\describe{
\item{\code{signals}}{Vector of length equal to \code{y} denoting presence or absence of peaks with three possible values: \code{0} (no peak), \code{1} (positive peak), \code{-1} (negative peak).}
\item{\code{avgFilter}}{Vector of length equal to \code{y} containing moving averages.}
\item{\code{stdFilter}}{Vector of length equal to \code{y} containing moving standard deviations.}
\item{\code{params}}{A vector of length 3 containing the used \code{lag}, \code{threshold}, and \code{influence} parameters passed to the function.}
}
}
\description{
Simple, robust peak picking algorithm based on calculating and thresholding z-scores.
}
\examples{
z_thres(lcqc::exlc1[,"Signal"], lag = 13, threshold = 4, influence = 0)

}
\references{
Brakel, J.P.G. van (2014). "Robust peak detection algorithm using z-scores". Stack Overflow. Available at: \url{https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data/22640362#22640362} (version: 2020-11-08).
}
\seealso{
\code{\link{z_optim}}
}
